% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accroissement.R
\name{accrois}
\alias{accrois}
\title{Fonction qui calcule de l'accroissement du diamètre des arbres pour un pas de simulation}
\usage{
accrois(Accrois, t, Para.acc, RandomAcc, Res)
}
\arguments{
\item{Accrois}{Un dataframe qui contient la liste des arbres et les variables nécessaire à
l'équation d'accroissement (au minimum Placette, ArbreID, DHPcm, GrEspece, vigu0, prod0, st_tot0, fact_red, ntrt, type_pe_Plac)}

\item{t}{La longueur du pas de simulation en annee (en annees).}

\item{Para.acc}{Un dataframe contenant les paramettres du module d'accroissement.}

\item{RandomAcc}{Un dataframe contenant les effets aléatoires du module d'accroissement.}

\item{Res}{Un dataframe d'une colonne contenant les erreurs residuelles du module d'accroissement.}
}
\value{
Retourne Accrois avec les colonnes DHPcm1, aam et pred_acc.
}
\description{
Fonction qui calcule de l'accroissement du diamètre des arbres pour un pas de simulation
}
