% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AttributionQualite.R
\name{AttribQualFct}
\alias{AttribQualFct}
\title{Fonction qui attribut la qualité ABCD aux tiges feuillues qui n'ont pas de valeurs de qualité.
La qualité est attribuée soit aux arbres dont la qualité n'a pas été évaluée ou
aux arbres qui en croissant atteignent le seuil de 23.1cm.}
\usage{
AttribQualFct(PlacSansQual, seed_value = NULL)
}
\arguments{
\item{PlacSansQual}{Un dataframe contenant une ligne par arbre avec au minimum les colonnes Placette, ArbreID, GrEspece, ABCD, DHPcm1, rid1}

\item{seed_value}{Optionnel. La valeur du seed pour la génération de nombres aléatoires. Généralement utilisé pour les tests de la fonction.}
}
\value{
Retourne PlacSansQual avec la colonne ABCD et PredQual
}
\description{
Fonction qui attribut la qualité ABCD aux tiges feuillues qui n'ont pas de valeurs de qualité.
La qualité est attribuée soit aux arbres dont la qualité n'a pas été évaluée ou
aux arbres qui en croissant atteignent le seuil de 23.1cm.
}
