% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AttributionVig0.R
\name{AttribVigu0}
\alias{AttribVigu0}
\title{Fonction qui extrait la vigueur à partir du code de vigueur au début de la simulation.
Si pas de code de vigueur, utilise les équations de conversion MSCR à vigueur}
\usage{
AttribVigu0(Data, Para.ConvMSCRVig)
}
\arguments{
\item{Data}{Un dataframe contenant une ligne par arbre avec au minimum les colonnes Vigueur, GrEspece, MSCR, DHPcm}

\item{Para.ConvMSCRVig}{Un dataframe  contenant les paramettres des équations
de conversion du classement MSCR en vigueur.}
}
\value{
Retourne Data avec la colonne vigu0 ("ViG" ou "NONVIG").
}
\description{
Fonction qui extrait la vigueur à partir du code de vigueur au début de la simulation.
Si pas de code de vigueur, utilise les équations de conversion MSCR à vigueur
}
