% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outils.R
\name{calcul_hdom}
\alias{calcul_hdom}
\title{Calculer la hauteur dominante de placettes}
\usage{
calcul_hdom(data, ht, nombre, superficie, group_by_vars = NULL)
}
\arguments{
\item{data}{Un dataframe dont les lignes sont des arbres individuels ou un groupe d'arbres, avec les colonnes hautm, nombre, superficie}

\item{ht}{Nom de la colonne dans \code{data} qui contient la hauteur de l'arbre en m.}

\item{nombre}{Nom de la colonne dans \code{data} qui contient le nombre d'arbres dans la placette que représente une ligne}

\item{superficie}{Nom de la colonne dans \code{data} qui contient la superficie de la placette en ha}

\item{group_by_vars}{Vecteur contenant le nom des colonnes dans \code{data} à utiliser pour grouper les lignes pour le calcul de la hauteur dominante. ex: c('Placette','Annee','GrEspece'). Facultatif.}
}
\value{
Un dataframe contenant, pour chaque placette, groupe d'espèces, année et itération,
la surface terrière, le volume marchand brut, le diamètre moyen quadratique et la hauteur dominante à l'ha.
}
\description{
Calculer la hauteur dominante de placettes
}
