% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outils.R
\name{calcul_var_dendro}
\alias{calcul_var_dendro}
\title{Calculer les variables dendrométriques de placettes}
\usage{
calcul_var_dendro(data, dhp, voldm3, nombre, superficie, group_by_vars = NULL)
}
\arguments{
\item{data}{Un dataframe dont les lignes sont des arbres individuels ou un groupe d'arbres, avec les colonnes dhp, voldm3, nombre, superficie}

\item{dhp}{Nom de la colonne dans \code{data} qui contient le DHP de l'arbre en cm}

\item{voldm3}{Nom de la colonne dans \code{data} qui contient le volume de l'arbre en dm3. Mettre des NA dans la colonne si le volume n'est pas disponible}

\item{nombre}{Nom de la colonne dans \code{data} qui contient le nombre d'arbres dans la placette que représente une ligne}

\item{superficie}{Nom de la colonne dans \code{data} qui contient la superficie de la placette en ha}

\item{group_by_vars}{Vecteur contenant le nom des colonnes dans \code{data} à utiliser pour grouper les lignes pour le calcul des variables dendrométriques. ex: c('Placette','Annee','GrEspece'). Facultatif.}
}
\value{
Un dataframe avec une ligne par combinaison des variables de \code{group_by_vars}, et les colonnes ST_m2ha (surface terrière m2/ha),
        Vol_m3ha (volume marchand brut m3/ha), Ti_HA (nombre d'arbres à l'ha), DQM_cm (diamètre moyen quadratique en cm).
}
\description{
Calculer les variables dendrométriques de placettes
}
