% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convertiseur_Eco.R
\name{ConvertisseurEco}
\alias{ConvertisseurEco}
\title{Fonction qui modifie la région écologique en vue de son utilisation par
le module de billonnage Petro régionalisé}
\usage{
ConvertisseurEco(data)
}
\arguments{
\item{data}{Un dataframe qui contient en ligne les arbres dont on veut prévoir
            les rendements en produit à l'aide du module de billonnage Petro
            régionalisé.

 @return Retourne un  dataframe qui contient en ligne les arbres dont on veut
         prévoir les rendements en produits à l'aide de Petro régionalisé avec le
         champ "eco" qui corespond à un groupement de régions écologiques
 @examples}
}
\description{
Fonction qui modifie la région écologique en vue de son utilisation par
le module de billonnage Petro régionalisé
}
