% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConversionMSCR.R
\name{ConvMSCRProd24}
\alias{ConvMSCRProd24}
\title{Fonction qui permet de convertir le classement MSCR Boulet-2007 en produits pour les arbres feuillus de
plus de 23,0 cm de diamètre selon le classement de Majcen-1990}
\usage{
ConvMSCRProd24(Data, Para.ConvMSCRProd24)
}
\arguments{
\item{Data}{Un dataframe contenant une ligne par arbre avec un champ MSCR copntenant le classement
MSCR de l'arbre. Les classements acceptés sont M, S, C, MS ou CR}

\item{Para.ConvMSCRProd1024}{Un dataframe  contenant les paramettres des équations de conversion du classement MSCR en vigueur}
}
\value{
Retourne un vecteur de probabilité que l'arbre soit de classe de produit sciage
}
\description{
Fonction qui permet de convertir le classement MSCR Boulet-2007 en produits pour les arbres feuillus de
plus de 23,0 cm de diamètre selon le classement de Majcen-1990
}
