% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaules_evolution.R
\name{count68HEG}
\alias{count68HEG}
\title{Fonction qui calcul le nombre de Gaules de hêtre à grandes feuilles classes de 6 ou 8 cm de diamètre lorsquelle sont présente.
Cette fonction utilise une variante des paramètre publiée par Rijal et al. 2023 Journal canadien de la recherche forestière.
Les prévisions sont basées sur un modèle de type Zero inflated}
\usage{
count68HEG(RecGaules, Ratio, Rec, RandomPlacGaules, Para.68_HEG)
}
\arguments{
\item{RecGaules}{Dataframe qui contient les variables lnNb_Gaules_68_Ess_Ha et lnNb_Gaules_24_Ess_Ha par Placette/GrEspece}

\item{Ratio}{Dataframe qui contient la variable Placette}

\item{Rec}{Dataframe avec la colonne Placette}

\item{RandomPlacGaules}{Dataframe contenant les effets aléatoires à l'échelle de la placette du module de gaules}

\item{Para.68_HEG}{Paramètres de l'équation de prévision du nombre de gaules de hêtre à grande feuille de 6 et 8 cm de diamètre}
}
\value{
Retourne un data d'une ligne par placette avec la colonne count_68HEG et xb_count_68HEG
}
\description{
Fonction qui calcul le nombre de Gaules de hêtre à grandes feuilles classes de 6 ou 8 cm de diamètre lorsquelle sont présente.
Cette fonction utilise une variante des paramètre publiée par Rijal et al. 2023 Journal canadien de la recherche forestière.
Les prévisions sont basées sur un modèle de type Zero inflated
}
