% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaules_evolution.R
\name{count68SAB}
\alias{count68SAB}
\title{Fonction qui calcul le nombre de Gaules de sapin baumier classes de 6 ou 8 cm de diamètre lorsquelle sont présente.
Cette fonction utilise une variante des paramètre publiée par Rijal et al. 2023 Journal canadien de la recherche forestière.
Les prévisions sont basées sur un modèle de type Zero inflated}
\usage{
count68SAB(RecGaules, Ratio, Rec, RandomPlacGaules, Para.68_SAB)
}
\arguments{
\item{RecGaules}{Dataframe qui contient la variable lnNb_Gaules_68_Ess_Ha par Placette/GrEspece}

\item{Ratio}{Dataframe qui contient les variables Nb_Gaules_Ha et lnNb_Gaules_Ess_Ha par Placette/GrEspece}

\item{Rec}{Dataframe qui contient les variables St_Ess_Ha, trt, t0_aj_ par placette/GrEspece}

\item{RandomPlacGaules}{Dataframe contenant les effets aléatoires à l'échelle de la placette du module de gaules}

\item{Para.68_SAB}{Paramètres de l'équation de prévision du nombre de gaules de sapin baumier de 6 et 8 cm de diamètre}
}
\value{
Retourne un data d'une ligne par placette avec la colonne count_68SAB et xb_count_68SAB
}
\description{
Fonction qui calcul le nombre de Gaules de sapin baumier classes de 6 ou 8 cm de diamètre lorsquelle sont présente.
Cette fonction utilise une variante des paramètre publiée par Rijal et al. 2023 Journal canadien de la recherche forestière.
Les prévisions sont basées sur un modèle de type Zero inflated
}
