% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cubage_arbres.R
\name{cubage_arbres}
\alias{cubage_arbres}
\title{Fonction qui estime la hauteur et le volume des arbres d'un fichier de résultats de simulation avec SaMARE}
\usage{
cubage_arbres(data, simul_data, NbIter, Horizon)
}
\arguments{
\item{data}{Dataframe d'intrant à la simulation avec les variables Placette, ArbreID, DHPcm, Sup_PE, Reg_Eco, Type_Eco, Altitude, Ptot, Tmoy
(dataframe préparé dans la fonction SimulSaMARE)}

\item{simul_data}{Dataframe des résultats de simulation avec SaMARE}

\item{NbIter}{Nombre d'iterations à effectuer (>=2 et un nombre pair).}

\item{Horizon}{Nombre de périodes de 5 ans sur à simuler, valeur de 1 à 12 (ex: 6 pour 30 ans de simulation).}
}
\value{
Retourne le dataframe Data avec les colonnes hauteur_pred et vol_dm3
}
\description{
Fonction qui estime la hauteur et le volume des arbres d'un fichier de résultats de simulation avec SaMARE
}
