% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outils.R
\name{dqm}
\alias{dqm}
\title{dqm
Calcul le diamètre quadratique moyen à partir de la la surface terrière et du nombre de tiges}
\usage{
dqm(stm2ha, nbha)
}
\arguments{
\item{stm2ha}{surface terrière en m2/ha}

\item{nbha}{Nombre d'arbres à l'ha}
}
\value{
diamètre quadratique moyen en cm
}
\description{
dqm
Calcul le diamètre quadratique moyen à partir de la la surface terrière et du nombre de tiges
}
