% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ess_groupe.R
\name{Ess_groupe}
\alias{Ess_groupe}
\title{set les paramettres}
\usage{
Ess_groupe(Data)
}
\arguments{
\item{Data}{data frame liste des donner de depart}
}
\description{
set les paramettres
}
\examples{
Ess_groupe(arbres)
}
