% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evolution_qualite.R
\name{EvolQual}
\alias{EvolQual}
\title{Fonction qui prévoit l'évolution de la classe de qualité des arbres.}
\usage{
EvolQual(PlacQual, Para.EvolQualTot, seed_value = NULL)
}
\arguments{
\item{PlacQual}{Un dataframe qui contient une liste d'arbres avec au minimum les colonnes Placette, ArbreID,
GrEspece ABCD Etat1 DHPcm1 type_pe_Plac vigu0 prod0 aam rid1 prec.}

\item{Para.EvolQualTot}{Un dataframe  contenant les paramètres du module d'évolution
de la qualité des arbres.}

\item{seed_value}{Optionnel. La valeur du seed pour la génération de nombres aléatoires. Généralement utilisé pour les tests de la fonction.}
}
\value{
PlacQual avec les colonnes ABCD, Intercept1, Intercept2, Intercept3.
}
\description{
Fonction qui prévoit l'évolution de la classe de qualité des arbres.
}
