% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaules_evolution.R
\name{gaules_68}
\alias{gaules_68}
\title{Fonction qui prévoit la proportion du nombre de gaules par espèce pour la période
de simulation suivante à partir des modèles de Rijal et al. 2023}
\usage{
gaules_68(
  Ratio,
  Rec,
  RecGaules,
  t,
  RandomPlacGaules,
  Para.68_ERS,
  Para.68_HEG,
  Para.68_BOJ,
  Para.68_SAB
)
}
\arguments{
\item{Ratio}{Dataframe qui contient les variables Nb_Gaules_Ha et lnNb_Gaules_Ess_Ha par Placette/GrEspece}

\item{Rec}{Dataframe qui contient les variables St_Ess_Ha, lnNb_Ess_Ha, trt, t0_aj_, altitude, latitude, dens_tot0, grwd par Placette/GrEspece}

\item{RecGaules}{Dataframe qui contient la variable lnNb_Gaules_68_Ess_Ha et lnNb_Gaules_24_Ess_Ha par Placette/GrEspece}

\item{t}{La longueur du pas de simulation en annees (en annees).}

\item{RandomPlacGaules}{Dataframe contenant les effets aléatoires à l'échelle de la placette du module de gaules}

\item{Para.68_ERS}{Paramètre du modèle du nombre de gaules 68 de ERS}

\item{Para.68_HEG}{Paramètre du modèle du nombre de gaules 68 de HEG}

\item{Para.68_BOJ}{Paramètre du modèle du nombre de gaules 68 de BOJ}

\item{Para.68_SAB}{Paramètre du modèle du nombre de gaules 68 de SAB}
}
\value{
Retourne un dataframe avec le nombre de gaules 24 et 68 par placette/espèce, colonnes
        Nb_Gaules_Ha, Nb_Gaules_Ess_Ha, Nb_Gaules_24_Ess_Ha, Nb_Gaules_68_Ess_Ha,
        lnNb_Gaules_Ess_Ha, lnNb_Gaules_24_Ess_Ha, lnNb_Gaules_68_Ess_Ha
}
\description{
Fonction qui prévoit la proportion du nombre de gaules par espèce pour la période
de simulation suivante à partir des modèles de Rijal et al. 2023
}
