% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GaulesTest2500m2}
\alias{GaulesTest2500m2}
\title{Fichier d'intrant des gaules}
\format{
## `GaulesTest2500m2`
A data frame with 18 lignes et 5 colonnes:
\describe{
  \item{Espece}{Code d'essence}
  \item{Placette}{Identifiant de la placette}
  \item{Nombre}{Nombre de gaules dans la classe de DHPcm et de Espece}
  \item{DHPcm}{Classe de DHP (cm)}
  \item{Sup_PE}{Superficie de la placette en ha}
}
}
\usage{
GaulesTest2500m2
}
\description{
une placette de 2500 m2 avec la liste de gaules
}
\examples{
GaulesTest2500m2
}
\keyword{datasets}
