% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mortalite.R
\name{mort}
\alias{mort}
\title{calcule de la mortalité}
\usage{
mort(Mort, trt, temp, type_pe_Plac, fact_red, t, Iterj, Para.mort)
}
\arguments{
\item{Mort}{liste des mort}

\item{trt}{Variable du peuplement residuel avec condition que si St >26 = TEM}

\item{temp}{temperature Variables climatiques de la placette}

\item{type_pe_Plac}{Type de placette ex: type0, type1, ou type2}

\item{fact_red}{facteur de Réduction de la mortalité}

\item{Iterj}{Iteration choisie}

\item{Para.mort}{Parametres de mortalité}
}
\description{
calcule de la mortalité
}
