% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaules_evolution.R
\name{nb_Gaules}
\alias{nb_Gaules}
\title{Fonction qui prévoit le nombre total de gaules à la fin du pas de simulation.}
\usage{
nb_Gaules(Rec, RecGaules, t, RandomPlacGaules, Para.nb_gaules)
}
\arguments{
\item{Rec}{Dataframe qui contient les variables St_Ess_Ha, lnNb_Ess_Ha, trt, t0_aj_, dens_tot0 par Placette/GrEspece}

\item{RecGaules}{Dataframe qui contient la variable Nb_Gaules_Ess_Ha par Placette/GrEspece}

\item{t}{La longueur du pas de simulation en annees (en annees).}

\item{RandomPlacGaules}{Dataframe contenant les effets aléatoires à l'échelle de la placette du module de gaules}

\item{Para.nb_gaules}{Paramètres de l'équation de la prévision du nombre total de gaules.}
}
\value{
Retourne un data d'une ligne par placette avec la colonne Nb_Gaules_Ha et xb_nb_gaules.
}
\description{
Fonction qui prévoit le nombre total de gaules à la fin du pas de simulation.
}
\details{
Cette fonction utilise une variante des paramètre publiée par Rijal et al. 2023 Journal canadien de la recherche forestière
}
