% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outils.R
\name{nbha}
\alias{nbha}
\title{nbha
Calcul le nombre d'arbres à l'ha que représente une ligne d'un data}
\usage{
nbha(nombre, superficie)
}
\arguments{
\item{nombre}{Nombre d'arbres dans la placette de superficie \code{superficie} que représente une ligne du data}

\item{superficie}{Superficie de la placette}
}
\value{
Nombre d'arbres à l'ha que représente une ligne d'un data
}
\description{
nbha
Calcul le nombre d'arbres à l'ha que représente une ligne d'un data
}
