% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaules_evolution.R
\name{pi68BOJ}
\alias{pi68BOJ}
\title{Fonction qui calcul la probabilité d'absence de Gaules de bouleau jaune classes
de 6 ou 8 cm de diamètre. Cette fonction utilise une variante des paramètre
publiée par Rijal et al. 2023 Journal canadien de la recherche forestière.
Les prévisions sont basées sur un modèle de type Zero inflated.}
\usage{
pi68BOJ(RecGaules, Ratio, Rec, RandomPlacGaules, Para.68_BOJ)
}
\arguments{
\item{RecGaules}{Dataframe qui contient la variable lnNb_Gaules_68_Ess_Ha par Placette/GrEspece}

\item{Ratio}{Dataframe qui contient les variables Nb_Gaules_Ha et lnNb_Gaules_Ess_Ha par Placette/GrEspece}

\item{Rec}{Dataframe qui contient les variables St_Ess_Ha, lnNb_Ess_Ha, trt, t0_aj_, altitude par Placette/GrEspece}

\item{RandomPlacGaules}{Dataframe contenant les effets aléatoires à l'échelle de la placette du module de gaules}

\item{Para.68_BOJ}{Paramètres de l'équation de prévision du nombre de gaules de bouleau jaune de 6 et 8 cm de diamètre.}
}
\value{
Retourne un data d'une ligne par placette avec la colonne pi_68BOJ et xb_pi_68BOJ.
}
\description{
Fonction qui calcul la probabilité d'absence de Gaules de bouleau jaune classes
de 6 ou 8 cm de diamètre. Cette fonction utilise une variante des paramètre
publiée par Rijal et al. 2023 Journal canadien de la recherche forestière.
Les prévisions sont basées sur un modèle de type Zero inflated.
}
