% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaules_68_ERS.R
\name{pi68ERS}
\alias{pi68ERS}
\title{Parametres nombre de gaules de 6 et 8 cm ERS}
\usage{
pi68ERS(RecGaules, Ratio, Iterj, RandomPlacGaules, Para.68_ERS)
}
\arguments{
\item{RecGaules}{Recrutement Gaules}

\item{Ratio}{Ratio Gaules}

\item{Iterj}{l'iteration souhaité}

\item{RandomPlacGaules}{placette de RandomGaules choisie}

\item{Para.68_ERS}{nombre de gaules de 6 et 8 cm ERS}
}
\description{
Parametres nombre de gaules de 6 et 8 cm ERS
}
