% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaules_68_HEG.R
\name{pi68HEG}
\alias{pi68HEG}
\title{Parametres nombre de gaules de 6 et 8 cm HEG}
\usage{
pi68HEG(RecGaules, Ratio, Rec, Iterj, RandomPlacGaules, Para.68_HEG)
}
\arguments{
\item{RecGaules}{Recrutement Gaules}

\item{Ratio}{Ratio Gaules}

\item{Rec}{Nombre de  Recrues}

\item{Iterj}{l'iteration souhaité}

\item{RandomPlacGaules}{placette de RandomGaules choisie}

\item{Para.68_HEG}{nombre de gaules de 6 et 8 cm HEG}
}
\description{
Parametres nombre de gaules de 6 et 8 cm HEG
}
