% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Produit.R
\name{produit}
\alias{produit}
\title{Fonction qui prévoit l'évolution de la classe de produit des tiges feuillus.}
\usage{
produit(Prod, Para.prod, RandomProd, seed_value = NULL)
}
\arguments{
\item{Prod}{Un dataframe qui contient les arbres avec au minimum les colonnes Placette, GrEspece, vigu0, type_pe_Plac, rid1, aam}

\item{Para.prod}{Un dataframe  contenant les parametres du module d'évolution
des produits des tiges feuillues.}

\item{RandomProd}{Un dataframe  contenant les effets aléatoires du module d'évolution
des produits des tiges feuillues.}

\item{seed_value}{Optionnel. La valeur du seed pour la génération de nombres aléatoires. Généralement utilisé pour les tests de la fonction.}
}
\value{
Retourne Prod avec les colonnes prod1 et prob_prod
}
\description{
Fonction qui prévoit l'évolution de la classe de produit des tiges feuillus.
}
