% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Random.R
\name{RandomPlacStep}
\alias{RandomPlacStep}
\title{Fonction qui simule les effets aléatoires à l'échelle de la placette
 et de la période de simulation pour les modules originaux du modèle SaMARE}
\usage{
RandomPlacStep(CovParms, Data, NbIter, NbPeriodes)
}
\arguments{
\item{CovParms}{Un dataframe contenant la variance des effets aléatoires en
lien avec leur module respectif.}

\item{Data}{Un dataframe contenant les arbres à simuler pour lesquels la
placette sera extraite afin d'y générer des effets aléatoire.}

\item{NbIter}{Nombre d'iterations à effectuer (>=2 et un nombre pair).}

\item{NbPeriodes}{Nombre de périodes pour lesquelles les effets aléatoires
seront simulés.}
}
\value{
Retourne une liste d'effets aléatoires à l'échelle de la placette et
         de la période de simulation.
}
\description{
Fonction qui simule les effets aléatoires à l'échelle de la placette
 et de la période de simulation pour les modules originaux du modèle SaMARE
}
