% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Random.R
\name{RandomPlacStepGaules}
\alias{RandomPlacStepGaules}
\title{Fonction qui simule les effets aléatoires à l'échelle de la placette
 pour les modules de recrutement basé sur les gaules et d'évolution des
 gaules de SaMARE}
\usage{
RandomPlacStepGaules(CovParms, Data, NbIter)
}
\arguments{
\item{CovParms}{Un dataframe contenant la variance des effets aléatoires en
lien avec leur module respectif}

\item{Data}{Un dataframe contenant les informations sur les gaules pour lesquels la
placette sera extraite afin d'y générer des effets aléatoire}

\item{NbIter}{Nombre d'itérations de la simulation}
}
\value{
Retourne une liste d'effets aléatoires à l'échelle de la placette
}
\description{
Fonction qui simule les effets aléatoires à l'échelle de la placette
 pour les modules de recrutement basé sur les gaules et d'évolution des
 gaules de SaMARE
}
\examples{
 #resultat <- RandomPlacStepGaules(CovParms, Data, NbIter)
print(resultat)

}
