% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaules_evolution.R
\name{ratio_Gaules}
\alias{ratio_Gaules}
\title{Fonction qui prévoit la proportion du nombre de gaules par espèce pour la période
de simulation suivante à partir des modèles de Rijal et al. 2023}
\usage{
ratio_Gaules(
  Ratio,
  Rec,
  RecGaules,
  t,
  RandomPlacGaules,
  Para.nb_gaules,
  Para.ratio_gaules
)
}
\arguments{
\item{Ratio}{Dataframe qui possède une ligne par groupe d'espèce dans lequel les prévisions de ratios seront rapportées.}

\item{Rec}{Dataframe qui contient les variables lnNb_Ess_Ha, longitude, latitude, st_tot0, dens_tot0, lnNb_Ess_Ha, prec, trt, t0_aj_ par Placette/GrEspece}

\item{RecGaules}{Dataframe qui contient la variable lnNb_Gaules_Ess_Ha par Placette/GrEspece}

\item{t}{La longueur du pas de simulation en annees (en annees).}

\item{RandomPlacGaules}{Un dataframe contenant les effets aléatoires à l'échelle de la placette du module des gaules.}

\item{Para.nb_gaules}{Paramètres de l'équation du nombre de gaules total}

\item{Para.ratio_gaules}{Paramètres de l'équation de la prévision du ratio du nombre de gaules par espèce.}
}
\value{
Retourne Ratio avec les colonnes Nb_Gaules_Ess_Ha et lnNb_Gaules_Ess_Ha
}
\description{
Fonction qui prévoit la proportion du nombre de gaules par espèce pour la période
de simulation suivante à partir des modèles de Rijal et al. 2023
}
