% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaules_evolution.R
\name{ratio_pi_Gaules}
\alias{ratio_pi_Gaules}
\title{Fonction qui qui prévoit la probabilité d'absence de gaules par espèce pour la période
de simulation suivante. Cette fonction est la première partie du modèle zero-inflated
de Rijal et al. 2023}
\usage{
ratio_pi_Gaules(Ratio, Rec, RecGaules, t, RandomPlacGaules, Para.ratio_gaules)
}
\arguments{
\item{Ratio}{Dataframe qui possède une ligne par groupe d'espèces dans lequel les prévisions de ratios seront rapportées.}

\item{Rec}{Dataframe qui contient les variables lnNb_Ess_Ha, longitude, latitude par Placette/GrEspece}

\item{RecGaules}{Dataframe qui contient les variables lnNb_Gaules_Ess_Ha et Ratio par Placette/GrEspece}

\item{t}{La longueur du pas de simulation en annees (en annees).}

\item{RandomPlacGaules}{Dataframe contenant les effets aléatoires à l'échelle de la placette du module de gaules
recrutement basé sur les gaules et du module d'évolution des gaules.}

\item{Para.ratio_gaules}{Paramètres de l'équation de la prévision du ratio du nombre de gaules par espèce.}
}
\value{
Retourne Ratio avec la colonne pred_pi_gaules et xb_pi_gaules.
}
\description{
Fonction qui qui prévoit la probabilité d'absence de gaules par espèce pour la période
de simulation suivante. Cette fonction est la première partie du modèle zero-inflated
de Rijal et al. 2023
}
