% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recrutement_Nb_avec_gaules.R
\name{rec_count_Gaules}
\alias{rec_count_Gaules}
\title{Fonction qui prévoie le nombre de recrues par groupe d'espèce lorsqu'il
y a présence de recrues du groupe d'espèce. Cette fonction corespond à la
deuxième portion de la fonction zero-inflated de prévision du nombre
de recrues de Rijal et al. 2023.}
\usage{
rec_count_Gaules(Rec, RecGaules, t, RandomPlacGaules, Para.rec_gaules)
}
\arguments{
\item{Rec}{Dataframe qui contient les variables st_tot0 et St_Ess_Ha et lnNb_Ess_Ha par Placette/GrEspece}

\item{RecGaules}{Dataframe qui contient les variables lnNb_Gaules_Ess_Ha et lnNb_Gaules_68_Ess_Ha par Placette/GrEspece}

\item{t}{La longueur du pas de simulation en annees (en annees).}

\item{RandomPlacGaules}{Un dataframe contenant les effets aléatoires àl'échelle de la placette du module de recrutement basé sur les gaules}

\item{Para.rec_gaules}{Paramètres de l'équation de prévivion du nombre de recrues utilisant le nombre de gaules.}
}
\value{
Retourne Rec avec la colonne predCount_nbrec et xb_count_rec
}
\description{
Fonction qui prévoie le nombre de recrues par groupe d'espèce lorsqu'il
y a présence de recrues du groupe d'espèce. Cette fonction corespond à la
deuxième portion de la fonction zero-inflated de prévision du nombre
de recrues de Rijal et al. 2023.
}
\details{
RecGaules et Rec doivent être triés de la même façon.
}
