% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recrutement_Nb_sans_gaules.R
\name{rec_delta}
\alias{rec_delta}
\title{Fonction qui prevoit le paramètre delta de la fonction Weibull qui est
la deuxième portion du modèles zero-inflated du nombre de recrues de SaMARE.}
\usage{
rec_delta(Rec, Para.rec_n)
}
\arguments{
\item{Rec}{Un dataframe dfe placettes/grespece avec les colonnes Placette, GrEspece, GrEssRec, st_tot0, ntrt, t0_aj_, logst_ess_1014}

\item{Para.rec_n}{Paramètres de l'équation du Delta modèle du nombre de recrues qui n'utilise
pas les recrues comme prédicteurs.}
}
\value{
Retourne Rec avec la colonne predDelta_nrec et xb_delta_rec.
}
\description{
Fonction qui prevoit le paramètre delta de la fonction Weibull qui est
la deuxième portion du modèles zero-inflated du nombre de recrues de SaMARE.
}
