% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recrutement_Dhp.R
\name{rec_dhp}
\alias{rec_dhp}
\title{Fonction qui estime le DHP des recrues}
\usage{
rec_dhp(
  RecSelect,
  t,
  Para.rec_dhp,
  RandomRecDhp,
  varRecDhp,
  theta,
  seed_value = NULL
)
}
\arguments{
\item{RecSelect}{Dataframe qui contient les prévisions du nombre de recrues par groupe d'espèces, avec au minimum les
colonnes Placette, GrEspece, st_tot0, dens_tot0, ntrt}

\item{t}{La longueur du pas de simulation en annees (en annees).}

\item{Para.rec_dhp}{Un dataframe  contenant les paramettres du module de prévision du DHP des recrues.}

\item{RandomRecDhp}{Effet aléatoires du modèle de dhp des recrues}

\item{varRecDhp}{Variance résiduelle du modèle de dhp des recrues}

\item{theta}{Paramètre de la fonction de variance résiduelle du modèle de dhp des recrues}

\item{seed_value}{Optionnel. La valeur du seed pour la génération de nombres aléatoires. Généralement utilisé pour les tests de la fonction.}
}
\value{
Retourne RecSelect avec la colonne DHPcm1 et pred_dhp
}
\description{
Fonction qui estime le DHP des recrues
}
