% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recrutement_Nb_sans_gaules.R
\name{rec_lambda}
\alias{rec_lambda}
\title{Fonction qui prevoit le paramètre lambda de la fonction Weibull qui est
la deuxième portion du modèles zero-inflated du nombre de recrues de SaMARE.}
\usage{
rec_lambda(Rec, t, Para.rec_n)
}
\arguments{
\item{Rec}{Un dataframe de placewtte/grespece avec les colonnes Placette, GrEspece, GrEssRec, type_pe_Plac, st_tot0, logst_ess_1014.}

\item{t}{La longueur du pas de simulation en annee (en annees).}

\item{Para.rec_n}{Paramètres de l'équation du lambda du modèle de nombre de recrues qui n'utilise
pas les recrues comme prédicteurs.}
}
\value{
Retourne Rec avec la colonne predLambda_nrec et xb_lambda_rec.
}
\description{
Fonction qui prevoit le paramètre lambda de la fonction Weibull qui est
la deuxième portion du modèles zero-inflated du nombre de recrues de SaMARE.
}
