% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recrutement_Nb_sans_gaules.R
\name{rec_n}
\alias{rec_n}
\title{Fonction qui prévoie le nombre de recrues par groupe d'espèces.
avec les fonctions rec_pi, rec_lambda et rec_delta}
\usage{
rec_n(Rec_data, t, Para.rec_n, RandomRec, seed_value = NULL)
}
\arguments{
\item{Rec_data}{Un dataframe des 10 groupes d'espèces pour chacune des placettes, avec les colonnes Placette, GrEspece,
et les variables nécessaires pour les fonctions rec_pi, rec_delta, rec_lambda}

\item{t}{La longueur du pas de simulation en annee (en annees).}

\item{Para.rec_n}{Paramètres des équations pour le nombre de recrues}

\item{RandomRec}{Effets aléatoires du module de probabilité de
recrues sans utiliser le nombre de gaules.}

\item{seed_value}{Optionnel. La valeur du seed pour la génération de nombres aléatoires. Généralement utilisé pour les tests de la fonction.}
}
\value{
Retourne un dataframe de placette/grespece avec la colonne NbRecrues.
}
\description{
Fonction qui prévoie le nombre de recrues par groupe d'espèces.
avec les fonctions rec_pi, rec_lambda et rec_delta
}
