% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recrutement_Nb_avec_gaules.R
\name{rec_n_Gaules}
\alias{rec_n_Gaules}
\title{Fonction qui prévoie le nombre de recrues par groupe d'espèces.
avec les fonctions rec_pi, rec_lambda et rec_delta}
\usage{
rec_n_Gaules(
  Rec,
  RecGaules,
  t,
  CovParmsGaules,
  RandomPlacGaules,
  Para.rec_gaules,
  seed_value = NULL
)
}
\arguments{
\item{Rec}{Dataframe qui contient les variables st_tot0 et St_Ess_Ha et lnNb_Ess_Ha par Placette/GrEspece}

\item{RecGaules}{Dataframe qui contient les variables lnNb_Gaules_Ess_Ha et lnNb_Gaules_68_Ess_Ha par Placette/GrEspece}

\item{t}{La longueur du pas de simulation en annees (en annees).}

\item{CovParmsGaules}{Dataframe contenant la variance des effets aléatoires des équations du module de recrutement basés sur l'information
provenant des gaules (modules 10 à 16).}

\item{RandomPlacGaules}{Un dataframe contenant les effets aléatoires à l'échelle de la placette du module de recrutement basé sur les gaules}

\item{Para.rec_gaules}{Paramètres de l'équation de prévivion du nombre de recrues utilisant le nombre de gaules}

\item{seed_value}{Optionnel. La valeur du seed pour la génération de nombres aléatoires. Généralement utilisé pour les tests de la fonction.}
}
\value{
Retourne un dataframe d'une ligne par placette/grespece avec la colonne NbRecrues
}
\description{
Fonction qui prévoie le nombre de recrues par groupe d'espèces.
avec les fonctions rec_pi, rec_lambda et rec_delta
}
\details{
RecGaules et Rec doivent être triés de la même façon.
}
