% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recrutement_Nb_sans_gaules.R
\name{rec_pi}
\alias{rec_pi}
\title{Fonction qui prévoie la probabilité de présence de recrues par grouoe d'espèce.
Cette fonction corespond à la première portion de la fonction zero-inflated
de prévision du nombre de recrues non basé sur les gaules.}
\usage{
rec_pi(Rec, t, Para.rec_n, RandomRec)
}
\arguments{
\item{Rec}{Un dataframe de placette/grespece avec les colonnes Placette, GrEspece, GrEssRec, st_tot0, ntrt, t0_aj_, type_pe_Plac, logst_ess_1014}

\item{t}{La longueur du pas de simulation en annee (en annees).}

\item{Para.rec_n}{Paramètres de l'équation de prévision du nombre de
recrues sans utiliser le nombre de gaules.}

\item{RandomRec}{Effets aléatoires du module de probabilité de
recrues sans utiliser le nombre de gaules.}
}
\value{
Retourne Rec avec la colonne predPi_nbrec et xb_prob_rec.
}
\description{
Fonction qui prévoie la probabilité de présence de recrues par grouoe d'espèce.
Cette fonction corespond à la première portion de la fonction zero-inflated
de prévision du nombre de recrues non basé sur les gaules.
}
