% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recrutement_Nb_avec_gaules.R
\name{rec_pi_Gaules}
\alias{rec_pi_Gaules}
\title{Fonction qui prévoie la probabilité d'absence de recrues par groupe d'espèce.
Cette fonction corespond à la première portion de la fonction zero-inflated
de prévision du nombre de recrues basé sur les gaules de Rijal et al. 2023.}
\usage{
rec_pi_Gaules(Rec, RecGaules, t, RandomPlacGaules, Para.rec_gaules)
}
\arguments{
\item{Rec}{Dataframe qui contient les variables st_tot0 et lnNb_Ess_Ha par Placette/GrEspece}

\item{RecGaules}{Dataframe qui contient les variables lnNb_Gaules_Ess_Ha et lnNb_Gaules_68_Ess_Ha par Placette/GrEspece}

\item{t}{La longueur du pas de simulation en annees (en annees).}

\item{RandomPlacGaules}{Un dataframe contenant les effets aléatoires à l'échelle de la placette du module de recrutement basé sur les gaules}

\item{Para.rec_gaules}{Paramètres de l'équation de prévivion du nombre de recrues utilisant le nombre de gaules.}
}
\value{
Retourne Rec avec la colonne predPi_nbrec et xb_prob_rec
}
\description{
Fonction qui prévoie la probabilité d'absence de recrues par groupe d'espèce.
Cette fonction corespond à la première portion de la fonction zero-inflated
de prévision du nombre de recrues basé sur les gaules de Rijal et al. 2023.
}
\details{
RecGaules et Rec doivent être triés de la même façon.
}
