% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recrutement_Produit.R
\name{rec_prod}
\alias{rec_prod}
\title{Fonction qui prevoit si une recrue a un potentiel de sciage ou de pate.}
\usage{
rec_prod(RecSelect, Para.rec_prod, RandomRecProd, seed_value = NULL)
}
\arguments{
\item{RecSelect}{Un dataframe qui contient une liste d'arbres avec les colonnes Placette, GrEspece, type_pe_Plac, rid1
par groupe d'espèce.}

\item{Para.rec_prod}{Paramètres de l'équation de la probabilité de présence
de potentiel sciage des recrues de groupe d'essence feuillus.}

\item{RandomRecProd}{Effets aléatoires du modèle de produit des recrues}

\item{seed_value}{Optionnel. La valeur du seed pour la génération de nombres aléatoires. Généralement utilisé pour les tests de la fonction.}
}
\value{
Retourne RecSelect avec les colonnes prob_prod_rec et prod1.
}
\description{
Fonction qui prevoit si une recrue a un potentiel de sciage ou de pate.
}
