% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmegaPara.R
\name{reconstruct}
\alias{reconstruct}
\title{Fonction qui reconstruit la matrice de variance-covariance à partir du
fichier Omega où chaque ligne correspond à une valeur de la matrice}
\usage{
reconstruct(m_lower, diag = TRUE, symmetric = TRUE)
}
\arguments{
\item{m_lower}{Vecteur de valeures de la matrice de variance-covariance}

\item{diag}{Champ TRUE/FALSE qui indique si on doit reconstruire la diagonale de la matrice}

\item{symmetric}{Champ TRUE/FALSE qui indique si la matrice est symetrique}
}
\value{
Retourne une matrice de variance-covariance symetrique
}
\description{
Fonction qui reconstruit la matrice de variance-covariance à partir du
fichier Omega où chaque ligne correspond à une valeur de la matrice
}
