% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trouver_noms_des_colonnes_absents.R
\name{renommer_les_colonnes}
\alias{renommer_les_colonnes}
\title{Renommer les colonnes  du fichier des arbres}
\usage{
renommer_les_colonnes(data)
}
\arguments{
\item{data}{Un dataframe, représentant le fichier des arbres, dont les colonnes doivent être renommées}
}
\description{
La fonction \code{renommer_les_colonnes} renomme les colonnes d'un dataframe.
Les noms doivent être ceux attendus avec seulement des différences au niveau des minuscules et majuscules.
}
\details{
La fonction suit les étapes suivantes :
\itemize{
  \item Définir un vecteur \code{ColOrdre} contenant les noms des colonnes souhaitées dans l'ordre désiré.
  \item Convertir les noms des colonnes existantes du dataframe en minuscules pour faciliter la correspondance.
  \item Convertir les noms des colonnes souhaitées en minuscules.
  \item Pour chaque nom de colonne souhaité, chercher sa correspondance parmi les noms des colonnes existantes.
  \item Si une correspondance est trouvée, renommer la colonne existante avec le nom souhaité.
  \item Retourner le dataframe avec les noms de colonnes mis à jour.
}
}
