% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaMARE.R
\name{SaMARE}
\alias{SaMARE}
\title{la simultation a l'échelle de placette}
\usage{
SaMARE(
  Random,
  RandomGaules,
  Data,
  Gaules,
  ListeIter,
  AnneeDep,
  Horizon,
  RecruesGaules,
  CovParms,
  CovParmsGaules,
  Para,
  ParaGaules,
  Omega,
  OmegaGaules
)
}
\arguments{
\item{Random}{Résultat de la fonction RandomPlacStep pour obtenir un effet aléatoire.}

\item{RandomGaules}{Résultat de la fonction RandPlacStepGaules pour obtenir un effet aléatoire pour les gaules .}

\item{Data}{Donne des arbres à simuler.}

\item{Gaules}{Donne des Gaules}

\item{ListeIter}{l'itération choisie}

\item{AnneeDep}{l'année de depart}

\item{Horizon}{Nombre de cycles de 5 ans souhaité}

\item{RecruesGaules}{Liste des nouvelles recrues}

\item{CovParms}{Paramètre pour un effet aléatoire}

\item{CovParmsGaules}{Paramètre pour un effet aléatoire pour gaules}
}
\description{
la simultation a l'échelle de placette
}
