% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sommaire_Classes_DHP.R
\name{Sommaire_Classes_DHP}
\alias{Sommaire_Classes_DHP}
\title{Fonction qui compile une sortie de SimulSAMARE par classe de DHP, par placette, groupe d'espèces, vigueur,
et fait la moyenne des itérations. La sortie présente donc un sommaire de la table de peuplement simulée
pour chacune des placettes.}
\usage{
Sommaire_Classes_DHP(SimulHtVol, simplifier = FALSE)
}
\arguments{
\item{SimulHtVol}{Un dataframe contenant les résultats de simulation pour chacune des
iterations du simulateur SaMARE. Typiquement un résultat retourné
par la fonction "SimulSaMARE".}

\item{simplifier}{Un booléen indiquant si les résultats doivent être simplifiés pour ne garder
que la première et la dernière année de la simulation. Par défaut, \code{FALSE}.}
}
\value{
Retourne un dataframe contenant le sommaire des itérations par placette, année
         groupe d'espèce, vigueur, classe de DHP.
}
\description{
Fonction qui compile une sortie de SimulSAMARE par classe de DHP, par placette, groupe d'espèces, vigueur,
et fait la moyenne des itérations. La sortie présente donc un sommaire de la table de peuplement simulée
pour chacune des placettes.
}
