% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sortie_Arbre_Samare.R
\name{SortieArbreSamare}
\alias{SortieArbreSamare}
\title{Fonction qui structure un dataframe de sortie dont chaque ligne correspond
à chacun des arbres par placette, par itération et par année.}
\usage{
SortieArbreSamare(SimulHtVol, simplifier = FALSE)
}
\arguments{
\item{SimulHtVol}{Un dataframe contenant les résultats de simulation pour chacune des
iterations du simulateur SaMARE. Typiquement un résultat retourné
par la fonction "SimulSaMARE".}

\item{simplifier}{Un booléen indiquant si les résultats doivent être simplifiés pour ne garder
que la première et la dernière année de la simulation. Par défaut, \code{FALSE}.}
}
\value{
Retourne un dataframe contenant l'ensemble des arbres pour chacune des
         placettes, années, itérations.
}
\description{
Fonction qui structure un dataframe de sortie dont chaque ligne correspond
à chacun des arbres par placette, par itération et par année.
}
