% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sortie_Billes_Samare_Colone.R
\name{SortieBillesSamareColone}
\alias{SortieBillesSamareColone}
\title{cette function prend en parametre un dafra qui a déja simuller et billoner}
\usage{
SortieBillesSamareColone(SimulHtVolBillon)
}
\description{
cette function prend en parametre un dafra qui a déja simuller et billoner
}
