% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sortie_Billes_Samare_Ligne.R
\name{SortieBillesSamareLigne}
\alias{SortieBillesSamareLigne}
\title{Fonction qui structure un dataframe de sortie pour lequels on rapporte en format ligne
pour les essences feuillues le volume prévu pour chacune des classe de produits
issues de la classification Petro.}
\usage{
SortieBillesSamareLigne(SimulHtVolBillon)
}
\arguments{
\item{SimulHtVolBillon}{Un dataframe de sortie de la fonction de billonnage.}
}
\value{
Retourne un dataframe contenant pour chaque arbre, placette, annee,
        iteration et classe de produits une ligne présentant le volume en dm3
        de chacune des classes de produit Petro des arbres de groupe d'essence
        feuillus. Les arbres d'essence résineuses sont présentés avec des
        valeures nulles de produits et de volume de produits.
}
\description{
Fonction qui structure un dataframe de sortie pour lequels on rapporte en format ligne
pour les essences feuillues le volume prévu pour chacune des classe de produits
issues de la classification Petro.
}
