% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sortie_Billonage.R
\name{SortieBillonage}
\alias{SortieBillonage}
\title{Fonction qui structure un dataframe de sortie dont chaque ligne correspond
à chacun des arbres par placette, par classe Petro, par itération et par année.}
\usage{
SortieBillonage(Data, Type)
}
\arguments{
\item{Data}{Un dataframe qui contient en ligne les arbres dont on veut prévoir
les rendements en produit à l'aide du module de billonnage Petro
régionalisé. sera applier sur "ERS", "BOJ", "ERR", "BOP", "HEG", "CHR"}

\item{Type}{"DHP" pour utiliser les équations régionalisées basées seulement sur le DHP
"ABCD" pour utiliser les équations régionalisées basées sur ABCD
"1234" pour utiliser les équations de 2015 basées sur 1234
"MSCR" pour utiliser les équations de 2015 basées sur MSCR
"DHP2015" pour utiliser les équations de 2015 basées seulement sur le DHP
"ABCD2015" pour utiliser les équations de 2015 basées sur ABCD}
}
\value{
Retourne un dataframe avec l'estimation du volume par classe de produit
         pour chacun des arbres feuillus.
}
\description{
Fonction qui structure un dataframe de sortie dont chaque ligne correspond
à chacun des arbres par placette, par classe Petro, par itération et par année.
}
