% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sortie_DendroIter_Samare.R
\name{SortieDendroIterSamare}
\alias{SortieDendroIterSamare}
\title{Structurer un dataframe de sortie pour les simulations SaMARE}
\usage{
SortieDendroIterSamare(SimulHtVol, simplifier = FALSE)
}
\arguments{
\item{SimulHtVol}{Un dataframe produit par la fonction \code{SimulSaMARE}.}

\item{simplifier}{Un booléen indiquant si les résultats doivent être simplifiés pour ne garder
que la première et la dernière année de la simulation. Par défaut, \code{FALSE}.}
}
\value{
Un dataframe contenant, pour chaque placette, groupe d'espèces, année et itération,
la surface terrière (ST_m2ha), le volume marchand brut (Vol_m3ha), le diamètre moyen quadratique (DQM_cm) et la hauteur dominante (Hdom_m).
}
\description{
La fonction \code{SortieDendroIterSamare} structure un dataframe de sortie pour lequel
on rapporte, pour chaque placette, année, groupe d'espèces et itération, le diamètre
quadratique moyen, la surface terrière, le volume et la hauteur dominante à l'ha.
Cette fonction prend en paramètre un dataframe produit par la fonction \code{SimulSaMARE}
}
