% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outils.R
\name{stm2ha}
\alias{stm2ha}
\title{stm2ha
Calcul la surface terrière en m2/ha que représente une ligne d'un data}
\usage{
stm2ha(dhpcm, nombre, superficie)
}
\arguments{
\item{dhpcm}{DHP de l'arbre en cm}

\item{nombre}{Nombre d'arbres dans la placette de superficie \code{superficie} que représente une ligne du data}

\item{superficie}{Superficie de la placette}
}
\value{
surface terrière en m2/ha que représente une ligne d'un data
}
\description{
stm2ha
Calcul la surface terrière en m2/ha que représente une ligne d'un data
}
