% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Test400m2Coupe}
\alias{Test400m2Coupe}
\title{Fichier d'intrant des arbres marchands avec coupe}
\format{
## `Test400m2Coupe`
A data frame with 18 lignes et 21 colonnes:
\describe{
  \item{Espece}{Code d'essence}
  \item{Placette}{Identifiant de la placette}
  \item{NoArbre}{Identifiant de l'arbre}
  \item{Nombre}{Nombre de tiges dans la classe de DHP=DHPcm et de l'essence=Espece}
  \item{DHPcm}{Classe de DHP (cm)}
  \item{Etat}{Code d'état de l'arbre}
  \item{Latitude}{Latitude de la placette}
  \item{Longitude}{Latitude de la placette}
  \item{Reg_Eco}{Code de la région écologique}
  \item{Type_Eco}{Code du type écologique}
  \item{Altitude}{Altitude (m)}
  \item{Tmoy}{Température annuelle moyenne  normale 30 ans}
  \item{Ptot}{Précipitation totale moyenne normale 30 ans}
  \item{GrwDays}{Nombre de jours de la saison de croissance moyenne normale 30 ans}
  \item{Vigueur}{Code de vigueur de l'arbre}
  \item{MSCR}{Code MSCR}
  \item{ABCD}{Code ABCD}
  \item{Sup_PE}{Superficie de la placette en ha}
  \item{Annee_Coupe}{Année de la dernière coupe partielle}
  \item{ntrt}{Nombre de coupes partielles}
  \item{Pente}{Classe de pente de la placette}
}
}
\usage{
Test400m2Coupe
}
\description{
Une placette de 400 m2 avec leur liste d'arbres
}
\examples{
Test400m2Coupe
}
\keyword{datasets}
