% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trouver noms des colonnes absents.R
\name{trouver_noms_absents}
\alias{trouver_noms_absents}
\title{Vérifier la présence des colonnes obligatoires dans le fichier d'arbres}
\usage{
trouver_noms_absents(Data)
}
\arguments{
\item{data}{Un dataframe représentant le fichier d'arbres.}
}
\value{
Une liste des noms des colonnes manquantes.
}
\description{
La fonction \code{trouver_noms_absents} vérifie si toutes les colonnes obligatoires
sont présentes dans un dataframe représentant le fichier d'arbres. Elle retourne une liste
des noms des colonnes manquantes, le cas échéant.
}
\examples{
\dontrun{
# Supposons que nous ayons un dataframe `data` avec les colonnes suivantes :
# "Placette", "NoArbre", "Espece", "Etat", "DHPcm", "Vigueur", "Nombre"

# Appel de la fonction
noms_absents <- trouver_noms_absents(data)

# Si les colonnes "Sup_PE", "Annee_Coupe", "Latitude", "Longitude", "Altitude", "Pente",
# "Reg_Eco", "Type_Eco", "MSCR", "ntrt", "ABCD" sont absentes, la fonction retournera :
# [1] "sup_pe" "annee_coupe" "latitude" "longitude" "altitude" "pente"
#     "reg_eco" "type_eco" "mscr" "ntrt" "abcd"
}

}
