% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation_colonne.R
\name{valide_Annee_Coupe}
\alias{valide_Annee_Coupe}
\title{Fonction pour vérifier que les valeurs saisies dans la colonne 'Annee_Coupe' sont correctes.}
\usage{
valide_Annee_Coupe(data)
}
\arguments{
\item{data}{fichier des arbres}
}
\value{
retourne vrai ou faux s'il détecte des erreurs.
}
\description{
Fonction pour vérifier que les valeurs saisies dans la colonne 'Annee_Coupe' sont correctes.
}
