% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation_colonne.R
\name{valide_Annee_depart}
\alias{valide_Annee_depart}
\title{Fonction pour vérifier l'année d'inventaire
Si Annee_invenaire n'est pas dans le fichier, la créer.
Vérifie que tous les arbres d'une placette ont la même année inventaire
Vérifie si la valeur de l'année d'inventaire n'est pas manquante}
\usage{
valide_Annee_depart(data)
}
\arguments{
\item{data}{fichier des arbres}
}
\description{
Fonction pour vérifier l'année d'inventaire
Si Annee_invenaire n'est pas dans le fichier, la créer.
Vérifie que tous les arbres d'une placette ont la même année inventaire
Vérifie si la valeur de l'année d'inventaire n'est pas manquante
}
