% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation_colonne.R
\name{valide_data}
\alias{valide_data}
\title{Répertoire des erreurs dans le fichier des arbres}
\usage{
valide_data(data)
}
\arguments{
\item{data}{Un dataframe représentant le fichier des arbres.}
}
\value{
Une liste des erreurs trouvées dans le fichier des arbres.
}
\description{
La fonction \code{valide_data} vérifie la validité des données dans un dataframe représentant
le fichier des arbres et retourne une liste des erreurs trouvées.
}
\examples{
\dontrun{
# Supposons que nous ayons un dataframe data représentant le fichier des arbres

# Appel de la fonction
# erreurs <- valide_data(data)

# La fonction retournera une liste des erreurs trouvées dans le dataframe erreurs
}

}
