% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation_colonne.R
\name{valide_data_gaules}
\alias{valide_data_gaules}
\title{Répertoire des erreurs dans le fichier des gaules}
\usage{
valide_data_gaules(data)
}
\arguments{
\item{data}{Un dataframe représentant le fichier des gaules.}
}
\value{
Une liste des erreurs trouvées dans le fichier des gaules.
}
\description{
La fonction \code{valide_data_gaules} vérifie la validité des données dans un dataframe représentant
le fichier des gaules et retourne une liste des erreurs trouvées.
}
\examples{
\dontrun{
# Supposons que nous ayons un dataframe data représentant le fichier des gaules

# Appel de la fonction
# erreurs <- valide_data_gaules(data)

# La fonction retournera une liste des erreurs trouvées dans le dataframe erreurs
}

}
