% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation_colonne.R
\name{valide_espece}
\alias{valide_espece}
\title{Fonction pour vérifier que les valeurs saisies dans la colonne 'Espece' sont correctes.}
\usage{
valide_espece(data)
}
\arguments{
\item{data}{fichier des arbres}
}
\value{
retourne vrai ou faux s'il détecte des erreurs.
}
\description{
Fonction pour vérifier que les valeurs saisies dans la colonne 'Espece' sont correctes.
}
\examples{

# Exemple avec un dataframe valide
valide_espece(data) # Devrait retourner TRUE

# Exemple avec un dataframe invalide (colonne 'Espece' manquante)
valide_espece(data) # Devrait retourner FALSE

# Exemple avec un dataframe invalide (valeurs non autorisées)
valide_espece(data) # Devrait retourner FALSE

# Exemple avec un dataframe invalide (valeurs NA)
valide_espece(data) # Devrait retourner FALSE
}
