% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation_colonne.R
\name{valide_Etat}
\alias{valide_Etat}
\title{Fonction pour vérifier que les valeurs saisies dans la colonne 'Etat' sont correctes.}
\usage{
valide_Etat(data)
}
\arguments{
\item{data}{fichier des arbres}
}
\value{
retourne vrai ou faux s'il détecte des erreurs.
}
\description{
Fonction pour vérifier que les valeurs saisies dans la colonne 'Etat' sont correctes.
}
