% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation_colonne.R
\name{valide_Nombre}
\alias{valide_Nombre}
\title{Fonction pour vérifier que les valeurs saisies dans la colonne 'nombre' sont correctes.}
\usage{
valide_Nombre(data)
}
\arguments{
\item{data}{fichier des arbres}
}
\value{
retourne vrai ou faux s'il détecte des erreurs.
 # Exemple avec un dataframe valide
valide_Nombre(data) # Devrait retourner TRUE

# Exemple avec un dataframe invalide (colonne 'Nombre' manquante)
valide_Nombre(data) # Devrait retourner FALSE

# Exemple avec un dataframe invalide (valeurs négatives)
valide_Nombre(data) # Devrait retourner FALSE

# Exemple avec un dataframe invalide (valeurs NA)
valide_Nombre(data) # Devrait retourner FALSE
}
\description{
Fonction pour vérifier que les valeurs saisies dans la colonne 'nombre' sont correctes.
}
