% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation_colonne.R
\name{valide_Nombre_gaules}
\alias{valide_Nombre_gaules}
\title{Cette fonction vérifie si la colonne 'Nombre' dans le fichier de données contient des valeurs valides pour chaque 'Placette'.
Elle s'assure que la colonne existe, qu'elle n'est pas vide, qu'elle ne contient pas de valeurs manquantes (NA), et que la somme des valeurs pour chaque 'Placette' est supérieure à zéro.}
\usage{
valide_Nombre_gaules(data)
}
\arguments{
\item{data}{Un dataframe contenant les données des arbres avec les colonnes 'Nombre' et 'Placette'.}
}
\value{
Retourne TRUE si les valeurs dans la colonne 'Nombre' sont valides pour chaque 'Placette', FALSE sinon.
}
\description{
Cette fonction vérifie si la colonne 'Nombre' dans le fichier de données contient des valeurs valides pour chaque 'Placette'.
Elle s'assure que la colonne existe, qu'elle n'est pas vide, qu'elle ne contient pas de valeurs manquantes (NA), et que la somme des valeurs pour chaque 'Placette' est supérieure à zéro.
}
\examples{

# Exemple avec un dataframe valide
valide_Nombre_gaules(data) # Devrait retourner TRUE

# Exemple avec un dataframe invalide (colonne 'Nombre' manquante)
valide_Nombre_gaules(data) # Devrait retourner FALSE

# Exemple avec un dataframe invalide (valeurs négatives)
valide_Nombre_gaules(data) # Devrait retourner FALSE

# Exemple avec un dataframe invalide (valeurs NA)
valide_Nombre_gaules(data) # Devrait retourner FALSE

# Exemple avec un dataframe invalide (somme de 'Nombre' par 'Placette' égale à zéro)
valide_Nombre_gaules(data) # Devrait retourner FALSE
}
