% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation_colonne.R
\name{valide_Pente}
\alias{valide_Pente}
\title{Fonction pour vérifier que les valeurs saisies dans la colonne 'Pente' sont correctes.}
\usage{
valide_Pente(data)
}
\arguments{
\item{data}{fichier des arbres}
}
\value{
retourne vrai ou faux s'il détecte des erreurs.
}
\description{
Fonction pour vérifier que les valeurs saisies dans la colonne 'Pente' sont correctes.
}
