% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation_colonne.R
\name{valide_Ptot}
\alias{valide_Ptot}
\title{Fonction pour vérifier que les valeurs saisies dans la colonne 'Ptot' sont correctes.}
\usage{
valide_Ptot(data)
}
\arguments{
\item{data}{fichier des arbres}
}
\value{
retourne vrai ou faux s'il détecte des erreurs.

# Exemple avec un dataframe valide
valide_Ptot(data) # Devrait retourner TRUE

# Exemple avec un dataframe invalide (colonne 'Ptot' manquante)
valide_Ptot(data) # Devrait retourner FALSE

# Exemple avec un dataframe invalide (valeurs en dehors de l'intervalle)
valide_Ptot(data) # Devrait retourner FALSE

# Exemple avec un dataframe invalide (valeurs NA)
valide_Ptot(data) # Devrait retourner FALSE

# Exemple avec un dataframe invalide (plusieurs valeurs différentes pour 'Ptot' par 'Placette')
valide_Ptot(data) # Devrait retourner FALSE
}
\description{
Fonction pour vérifier que les valeurs saisies dans la colonne 'Ptot' sont correctes.
}
