% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation_colonne.R
\name{valide_Vigueur}
\alias{valide_Vigueur}
\title{Fonction pour vérifier que les valeurs saisies dans la colonne 'Vigueur' sont correctes.}
\usage{
valide_Vigueur(data)
}
\arguments{
\item{data}{fichier des arbres}
}
\value{
retourne vrai ou faux s'il détecte des erreurs.
}
\description{
Fonction pour vérifier que les valeurs saisies dans la colonne 'Vigueur' sont correctes.
}
\examples{

# Exemple avec un dataframe valide
valide_Vigueur(data) # Devrait retourner TRUE

# Exemple avec un dataframe invalide (valeurs non autorisées dans 'Vigueur')
valide_Vigueur(data) # Devrait retourner FALSE

# Exemple avec un dataframe invalide (valeurs NA sans MSCR correct)
valide_Vigueur(data) # Devrait retourner FALSE
}
