% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trouver_noms_des_colonnes_absents.R
\name{verifier_variable_meteo}
\alias{verifier_variable_meteo}
\title{Fonction qui vérifie si les variables climatiques sont présentes. Si elle sont absentes, elles
sont estimée à l'aide du package extract_map.}
\usage{
verifier_variable_meteo(data)
}
\arguments{
\item{data}{Un dataframe contenant la liste d'arbres à simuler}
}
\value{
Retourne la liste d'arbre initiale avec les données météo ajoutée si
       elles sont absentes
}
\description{
Fonction qui vérifie si les variables climatiques sont présentes. Si elle sont absentes, elles
sont estimée à l'aide du package extract_map.
}
