% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trouver_noms_des_colonnes_absents.R
\name{verifier_variable_station}
\alias{verifier_variable_station}
\title{Fonction qui vérifie si la pente est dans le fichier.
Si elle est absente, elle est estimée à l'aide du package ExtractMap}
\usage{
verifier_variable_station(data)
}
\arguments{
\item{data}{Un dataframe contenant la liste d'arbres à simuler.}
}
\value{
Retourne la liste d'arbres initiale avec les données de station ajoutées si
       elles sont absentes.
}
\description{
Fonction qui vérifie si la pente est dans le fichier.
Si elle est absente, elle est estimée à l'aide du package ExtractMap
}
