% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vigueur.R
\name{vig}
\alias{vig}
\title{Fonction qui prévoit l'évolution de la classe de vigueur des arbres.
La vigueur peut prendre 2 valeurs soit "vigoureux" ou "non-vigoureux"}
\usage{
vig(Vig, Para.vig, RandomVig, seed_value = NULL)
}
\arguments{
\item{Vig}{Un dataframe qui contient une liste d'arbres avec les colonnes Placette, GrEspece, DHPcm, vigu0, type_pe_Plac, rid1, aam}

\item{Para.vig}{Un dataframe  contenant les parametres du module d'évolution
de la vigueur des arbres.}

\item{RandomVig}{Un dataframe  contenant les effets aléatoires du module d'évolution
de la vigueur des arbres.}

\item{seed_value}{Optionnel. La valeur du seed pour la génération de nombres aléatoires. Généralement utilisé pour les tests de la fonction.}
}
\value{
Retourne Vig avec les colonnes vigu1 et prob_vig et xb_vig.
}
\description{
Fonction qui prévoit l'évolution de la classe de vigueur des arbres.
La vigueur peut prendre 2 valeurs soit "vigoureux" ou "non-vigoureux"
}
