% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outils.R
\name{volm3ha}
\alias{volm3ha}
\title{volm3ha
Calcul le volume en m3/ha que représente une ligne d'un data}
\usage{
volm3ha(voldm3, nombre, superficie)
}
\arguments{
\item{voldm3}{volume de l'arbre en dm3}

\item{nombre}{Nombre d'arbres dans la placette de superficie \code{superficie} que représente une ligne du data}

\item{superficie}{Superficie de la placette}
}
\value{
surface terrière en m2/ha que représente une ligne d'un data
}
\description{
volm3ha
Calcul le volume en m3/ha que représente une ligne d'un data
}
