% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accroissement.R
\name{accrois}
\alias{accrois}
\title{calcule l'Accroissement}
\usage{
accrois(Accrois, st_tot0, t, fact_red, ntrt, type_pe_Plac, Iterj, Para.acc)
}
\arguments{
\item{Accrois}{Data frame   contient les placette}

\item{st_tot0}{Nombre variable echelle placette}

\item{t}{Nombre  période de temps entre chaque simulation}

\item{fact_red}{Nombre  facteur de Réduction de la mortalité}

\item{ntrt}{Nombre  le nombre de traitements}

\item{type_pe_Plac}{Character  Type de placette}

\item{Iterj}{Nombre Iteration}

\item{Para.acc}{Data frame  paramettres d'accroissement}
}
\description{
calcule l'Accroissement
}
\examples{
  accrois(Accrois,21.09462,5,0,0,"type1",1,Para.acc)
}
