% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AttributionProd0.R
\name{AttribProd0}
\alias{AttribProd0}
\title{Fonction qui attribut les produits au début de la simulation. La fonction
cherche dabord si la tige est résineuse ensuite si le classement
vigueur 1,2,3,4,5,6 est fourni sinon elle cherche le classement MSCR pour
attribuer le produit sinon il classe l'arbre comme pate.}
\usage{
AttribProd0(Data, Para.ConvMSCRProd1024, Para.ConvMSCRProd24)
}
\arguments{
\item{Data}{Un dataframe contenant une ligne par arbre pour lesquels on
veut attribuer le vigueur.}

\item{Para.ConvMSCRProd1024}{Un dataframe  contenant les paramettres des équations
de conversion du classement MSCR en produits pour les arbres de moins de 23.1 cm.}

\item{Para.ConvMSCRProd24}{Un dataframe  contenant les paramettres des équations
de conversion du classement MSCR en produits pour les arbres de plus de 23 cm.}
}
\value{
Retourne la classe de produit de l'arbre.
}
\description{
Fonction qui attribut les produits au début de la simulation. La fonction
cherche dabord si la tige est résineuse ensuite si le classement
vigueur 1,2,3,4,5,6 est fourni sinon elle cherche le classement MSCR pour
attribuer le produit sinon il classe l'arbre comme pate.
}
\examples{

#l'attribution des produits
resultat <- AttribProd0(Data,Para.ConvMSCRProd1024,Para.ConvMSCRProd24)
print(resultat)

}
