% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AttributionVig0.R
\name{AttribVigu0}
\alias{AttribVigu0}
\title{Fonction qui attribut la vigueur au début de la simulation. La fonction
cherche dabord si le classement vigueur 1,2,3,4,5,6 est fourni sinon
elle cherche le classement MSCR pour attribuer la vigeur sinon il classe l'arbre
comme vigoureux.}
\usage{
AttribVigu0(Data, Para.ConvMSCRVig)
}
\arguments{
\item{Data}{Un dataframe contenant une ligne par arbre pour lesquels on
veut attribuer le vigueur.}

\item{Para.ConvMSCRVig}{Un dataframe  contenant les paramettres des équations
de conversion du classement MSCR en vigueur.}
}
\value{
Retourne la vigueur de l'arbre vigoureux ("ViG")ou non vigoureux ("NONVIG").
}
\description{
Fonction qui attribut la vigueur au début de la simulation. La fonction
cherche dabord si le classement vigueur 1,2,3,4,5,6 est fourni sinon
elle cherche le classement MSCR pour attribuer la vigeur sinon il classe l'arbre
comme vigoureux.
}
\examples{

#l'attribution de la vigueur
resultat <- AttribVigu0(Data,Para.ConvMSCRVig)
print(resultat)

}
