% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Billonnage.R
\name{Billonnage}
\alias{Billonnage}
\title{calcule du billonnage}
\usage{
Billonnage(arbres)
}
\arguments{
\item{arbres}{data frame liste des arbes a billoner}
}
\description{
calcule du billonnage
}
\examples{
Billonnage(arbres)
}
