% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaules_evolution.R
\name{count68BOJ}
\alias{count68BOJ}
\title{Fonction qui calcul le nombre de Gaules de bouleau jaune classes de 6 ou 8 cm de diamètre lorsquelle sont présente.
Cette fonction utilise une variante des paramètre publiée par Rijal et al. 2023 Journal canadien de la recherche forestière.
Les prévisions sont basées sur un modèle de type Zero inflated}
\usage{
count68BOJ(RecGaules, Ratio, Rec, t, RandomPlacGaules, Para.68_BOJ)
}
\arguments{
\item{RecGaules}{Dataframe qui contient la variable lnNb_Gaules_68_Ess_Ha par Placette/GrEspece}

\item{Ratio}{Dataframe qui contient la variable lnNb_Gaules_Ess_Ha par Placette/GrEspece}

\item{Rec}{Dataframe qui contient les variables à l'échelle de la placette trt, t0_aj_, latitude par Placette}

\item{t}{Longueur d'un pas de simulation}

\item{RandomPlacGaules}{Dataframe contenant les effets aléatoires à l'échelle de la placette du module de gaules}

\item{Para.68_BOJ}{Paramètre de l'équation de prévision du nombre de gaules de bouleau jaune de 6 et 8 cm de diamètre}
}
\value{
Retourne un data d'une ligne par placette avec la colonne count_68BOJ et xb_count_68BOJ
}
\description{
Fonction qui calcul le nombre de Gaules de bouleau jaune classes de 6 ou 8 cm de diamètre lorsquelle sont présente.
Cette fonction utilise une variante des paramètre publiée par Rijal et al. 2023 Journal canadien de la recherche forestière.
Les prévisions sont basées sur un modèle de type Zero inflated
}
