% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaules_evolution.R
\name{count68ERS}
\alias{count68ERS}
\title{Fonction qui calcul le nombre de Gaules d'érable à sucre classes de 6 ou 8 cm de diamètre lorsquelle sont présente.
Cette fonction utilise une variante des paramètre publiée par Rijal et al. 2023 Journal canadien de la recherche forestière.
Les prévisions sont basées sur un modèle de type Zero inflated}
\usage{
count68ERS(RecGaules, Ratio, Rec, RandomPlacGaules, Para.68_ERS)
}
\arguments{
\item{RecGaules}{Dataframe qui contient la variable lnNb_Gaules_68_Ess_Ha par Placette/GrEspece}

\item{Ratio}{Dataframe qui contient les variables Nb_Gaules_Ha et lnNb_Gaules_Ess_Ha par Placette/GrEspece}

\item{Rec}{Dataframe qui contient les variables à l'échelle de la placette dens_tot0, grwd par placette}

\item{RandomPlacGaules}{Dataframe contenant les effets aléatoires à l'échelle de la placette du module de gaules}

\item{Para.68_ERS}{Paramètres de l'équation de prévision du nombre de gaules d'érable à sucre de 6 et 8 cm de diamètre}
}
\value{
Retourne un data d'une ligne par placette avec la colonne count_68ERS et xb_count_68ERS
}
\description{
Fonction qui calcul le nombre de Gaules d'érable à sucre classes de 6 ou 8 cm de diamètre lorsquelle sont présente.
Cette fonction utilise une variante des paramètre publiée par Rijal et al. 2023 Journal canadien de la recherche forestière.
Les prévisions sont basées sur un modèle de type Zero inflated
}
