% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphiques.R
\name{Graph}
\alias{Graph}
\title{Fonction qui prépare les graphiques à insérer dans l'interface utilisateur
cette function prend en parametre une sortie de la fonction SimulSaMARE}
\usage{
Graph(SimulHtVol, Espece = "TOT", Variable = "ST_HA")
}
\arguments{
\item{SimulHtVol}{Un dataframe contenant les résultats de simulation pour chacune des
iterations du simulateur SaMARE. Typiquement un résultat retourné
par la fonction "SimulSaMARE".}

\item{Espece}{Un code de trois lettre majuscule présentant l'espèce pour laquelle on veut faire
les graphiques. Le code d'espèce doit être celui d'un des groupes d'espèce de SaMARE
ou le code "TOT" pour l'ensemble des espèces.}
}
\value{
Retourne une liste de deux graphiques, le premier montrant l'évolution d'une variables
         dendrométrique et le second la distribution en diamètre au début et à la fin de la simulation
         moyen quadratique et la hauteur dominante.
}
\description{
Fonction qui prépare les graphiques à insérer dans l'interface utilisateur
cette function prend en parametre une sortie de la fonction SimulSaMARE
}
