% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mortalite.R
\name{mort}
\alias{mort}
\title{Fonction qui prévoit la probabilité de mortalité de chacun des arbres
durant un pas de simulation.}
\usage{
mort(Mort, t, MCH, Para.mort, RandomMort, seed_value = NULL)
}
\arguments{
\item{Mort}{Un dataframe qui contient la liste des arbres avec au minimum les colonnes Placette, GrEspece, DHPcm,
vigu0, prod0, trt, temp, type_pe_Plac, fact_red}

\item{t}{La longueur du pas de simulation en annee (en annees).}

\item{MCH}{Binaire. 1 = indique la présence de la maladie corticale du hêtre durant tout l'horizon de simulation.
0 = absence de la maladie (par défaut).}

\item{Para.mort}{Un dataframe  contenant les paramettres du module de mortalité.}

\item{RandomMort}{Un dataframe contenant les effets aléatoires du module de mortalité}

\item{seed_value}{Optionnel. La valeur du seed pour la génération de nombres aléatoires. Généralement utilisé pour les tests de la fonction.}
}
\value{
Retourne Mort avec les colonnes prob_mort , xb_mort et Etat1
}
\description{
Fonction qui prévoit la probabilité de mortalité de chacun des arbres
durant un pas de simulation.
}
