% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mortalite.R
\name{mort}
\alias{mort}
\title{Fonction qui prévoit la probabilité de mortalité de chacun des arbres
durant un pas de simulation.}
\usage{
mort(Mort, trt, temp, type_pe_Plac, fact_red, t, Iterj, MCH, Para.mort)
}
\arguments{
\item{Mort}{Un dataframe qui contient la liste des arbres pour lesquels
la probabilité de mortalité doit être évaluée ainsi que les
caractéristiques de ses arbres qui seront utilisées
pour en prévoir le risque de mortalité.}

\item{trt}{Variable distinguant les peuplements traités des témoins, si St >26 = TEM.}

\item{temp}{Température annuelle moyenne de la placette.}

\item{type_pe_Plac}{Variable indicatrice de la taille de la placette soit
400 m2, entre 2500 et 5000 m2 inclusivement ou
une autre dimension.}

\item{fact_red}{Facteur de correction appliqué lorsqu'une coupe partielle a
été effectuée 3 ans ou moins avant la prévision.}

\item{t}{La longueur du pas de simulation en annee (en annees).}

\item{Iterj}{Itération en cours.}

\item{MCH}{Variable prenant la veleur de 1 en présence de maladie corticale du hêtre dans
la placette et 0 lorsque la maladie est absente. Lorsque la maladie corticale
est présente,la probabilité de mortalié des hêtres est estimée avec
l'équation de l'avis technique AT-SSRF 20 de la Direction de la recherche forestière.}

\item{Para.mort}{Un dataframe  contenant les paramettres du module de mortalité.}
}
\value{
Retourne le prédicteur linéaire de l'équation de la prévision du
        risque de mortalité. Les valeurs présites sont faites sans effets
        aléatoires, ceux-ci sont ajoutés dans la fonction SaMARE avant
        de convertir le prédicteur linéaire en probabilité de mortalité.
}
\description{
Fonction qui prévoit la probabilité de mortalité de chacun des arbres
durant un pas de simulation.
}
