% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaules_Nb.R
\name{nb_Gaules}
\alias{nb_Gaules}
\title{Fonction qui prévoit le nombre total de gaules à la fin du pas de simulation.
Cette fonction utilise une variante des paramètre publiée par Rijal et al. 2023 Journal canadien de la recherche forestière}
\usage{
nb_Gaules(
  Rec,
  RecGaules,
  t,
  st_tot0,
  altitude,
  latitude,
  trt,
  t0_aj_,
  longitude,
  temp,
  pente,
  Iterj,
  RandomPlacGaules,
  Para.nb_gaules
)
}
\arguments{
\item{Rec}{Un dataframe qui contient la prévision du nombre de recrues par
groupes d'espèces.}

\item{RecGaules}{Dataframe qui contient les information sur la distribution
des gaules dans la placette.}

\item{t}{La longueur du pas de simulation en annees (en annees).}

\item{st_tot0}{Surface terrière marchande (DHP >9.0cm) de la placette
au début du pas de simulation.}

\item{altitude}{Altitude de la placette.}

\item{latitude}{Latitude de la placette en degrés décimaux.}

\item{trt}{Variable distinguant les peuplements traités des témoins, si St >26 = TEM}

\item{t0_aj_}{Temps écoulé depuis la dernière coupe partielle}

\item{longitude}{Longitude de la placette en degrés décimaux.}

\item{temp}{Température annuelle moyenne de la placette.}

\item{pente}{Inclinaison de la pente  en %.}

\item{Iterj}{Itération en cours.}

\item{RandomPlacGaules}{Un dataframe contenant les effets aléatoires à
 l'échelle de la placette du module de
 recrutement basé sur les gaules et du module
d'évolution des gaules.}

\item{Para.nb_gaules}{Paramètres de l'équation de la prévision du nombre total de gaules.}
}
\value{
Retourne une prévision du nombre total de gaules à la fin du pas de simulation.
}
\description{
Fonction qui prévoit le nombre total de gaules à la fin du pas de simulation.
Cette fonction utilise une variante des paramètre publiée par Rijal et al. 2023 Journal canadien de la recherche forestière
}
