% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parametres_Simul_SaMARE.R
\name{ParametresSimulSaMARE}
\alias{ParametresSimulSaMARE}
\title{Fonction qui lit les paramètres nécessaire au simulateur SaMARE}
\usage{
ParametresSimulSaMARE()
}
\value{
Retourne une liste de dataframes qui contiennent les paramètres
        nécessaires à la simulation
}
\description{
Fonction qui lit les paramètres nécessaire au simulateur SaMARE
}
