% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmegaPara.R
\name{ParaOmega}
\alias{ParaOmega}
\title{Fonction qui ajuste les paramètre des chaque itération en fonction de leur
variance-covariance.}
\usage{
ParaOmega(ModuleID, ParaOri, ParaIter, Omega, NbIter)
}
\arguments{
\item{ModuleID}{Numéro du module pour lequel on veut simuler les paramètres.}

\item{ParaOri}{Dataframe qui contient les paramètres de base que l'on veut simuler.}

\item{ParaIter}{Copie du dataframe ParaOri répliqué n=NbIter fois.}

\item{Omega}{Dataframe qui contient les valeurs des matrices de variance-covariance.}

\item{NbIter}{Nombre d'iterations, corespond au nombre de version des
paramètres que l'on veut obtenir.}
}
\description{
Fonction qui ajuste les paramètre des chaque itération en fonction de leur
variance-covariance.
}
\examples{
# ajuste les paramètre des chaque itération
resultat <- ParaOmega(ModuleID,ParaOri,ParaIter,Omega,NbIter)
print(resultat)

}
