% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaules_68_BOJ.R
\name{pi68BOJ}
\alias{pi68BOJ}
\title{Parametres nombre de gaules de 6 et 8 cm BOJ}
\usage{
pi68BOJ(
  RecGaules,
  Ratio,
  Rec,
  trt,
  t0_aj_,
  altitude,
  Iterj,
  RandomPlacGaules,
  Para.68_BOJ
)
}
\arguments{
\item{RecGaules}{Recrutement Gaules}

\item{Ratio}{Ratio Gaules}

\item{Rec}{Nombre de  Recrues}

\item{trt}{Variable du peuplement residuel avec condition que si St >26 = TEM}

\item{t0_aj_}{Temps depuis coupe}

\item{altitude}{l'altitude une Variables de classification écologiques}

\item{Iterj}{l'iteration souhaité}

\item{RandomPlacGaules}{placette de RandomGaules choisie}

\item{Para.68_BOJ}{nombre de gaules de 6 et 8 cm BOJ}
}
\description{
Parametres nombre de gaules de 6 et 8 cm BOJ
}
