% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaules_68_SAB.R
\name{pi68SAB}
\alias{pi68SAB}
\title{Parametres nombre de gaules de 6 et 8 cm SAB}
\usage{
pi68SAB(RecGaules, Ratio, dens_tot0, Iterj, RandomPlacGaules, Para.68_SAB)
}
\arguments{
\item{RecGaules}{Recrutement Gaules}

\item{Ratio}{Ratio Gaules}

\item{dens_tot0}{densité total}

\item{Iterj}{l'iteration souhaité}

\item{RandomPlacGaules}{placette de RandomGaules choisie}

\item{Para.68_BOJ}{nombre de gaules de 6 et 8 cm SAB}
}
\description{
Parametres nombre de gaules de 6 et 8 cm SAB
}
