% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Produit.R
\name{produit}
\alias{produit}
\title{Fonction qui prévoit l'évolution de la classe de produit des tiges feuillus.
 Produit qui est contenu dans les classes "sciage" ou "pate".}
\usage{
produit(Prod, type_pe_Plac, rid1, Iterj, Para.prod)
}
\arguments{
\item{Prod}{Un dataframe qui contient les arbres pour lesquels on veut
prévoir l'évolution de la classe de produit des arbres feuillus.}

\item{type_pe_Plac}{Variable indicatrice de la taille de la placette soit
400 m2, soit entre 2500 et 5000 m2 inclusivement ou
une autre dimension.}

\item{rid1}{Variable de groupement de variables écologiques.}

\item{Iterj}{Itération en cours.}

\item{Para.prod}{Un dataframe  contenant les parametres du module d'évolution
des produits des tiges feuillues.}
}
\value{
Retourne le prédicteur linéaire de l'équation de la probabilité que
        la tige contienne un produit de type "sciage".
}
\description{
Fonction qui prévoit l'évolution de la classe de produit des tiges feuillus.
 Produit qui est contenu dans les classes "sciage" ou "pate".
}
