% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaules_evolution.R
\name{ratio_count_Gaules}
\alias{ratio_count_Gaules}
\title{Fonction qui prévoit le nombre de gaules par espèce quand l'espece est présente pour la période
de simulation suivante. Cette fonction est la deuxième partie du modèle zero-inflated
de Rijal et al. 2023.}
\usage{
ratio_count_Gaules(
  Ratio,
  Rec,
  RecGaules,
  t,
  RandomPlacGaules,
  Para.ratio_gaules
)
}
\arguments{
\item{Ratio}{Dataframe qui possède une ligne par groupe d'espèce dans lequel les prévisions de ratios seront rapportées.}

\item{Rec}{Dataframe qui contient les variables st_tot0, dens_tot0, lnNb_Ess_Ha, prec, trt, t0_aj_, longitude Placette/GrEspece}

\item{RecGaules}{Dataframe qui contient les variables lnNb_Gaules_Ess_Ha et Ratio par Placette/GrEspece}

\item{t}{La longueur du pas de simulation en annees (en annees).}

\item{RandomPlacGaules}{Dataframe contenant les effets aléatoires à l'échelle de la placette du module de gaules}

\item{Para.ratio_gaules}{Paramètres de l'équation de la prévision du ratio du nombre de gaules par espèce.}
}
\value{
Retourne Ratio avec la colonne pred_count_gaules et xb_count_gaules.
}
\description{
Fonction qui prévoit le nombre de gaules par espèce quand l'espece est présente pour la période
de simulation suivante. Cette fonction est la deuxième partie du modèle zero-inflated
de Rijal et al. 2023.
}
