% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ratio_pi_Gaules.R
\name{ratio_pi_Gaules}
\alias{ratio_pi_Gaules}
\title{Fonction qui qui prévoit la proportion du nombre de gaules par espèce pour la période
de simulation suivante. Cette fonction est la première partie du modèle zero-inflated
de Rijal et al. 2023 et prévoit une probabilité d'absence de gaules par
groupe d'espèce.}
\usage{
ratio_pi_Gaules(
  Ratio,
  Rec,
  RecGaules,
  t,
  st_tot0,
  latitude,
  longitude,
  Iterj,
  RandomPlacGaules,
  Para.ratio_gaules
)
}
\arguments{
\item{Ratio}{Dataframe qui possède une ligne par groupe d'espèces dans lequel
les prévisions de ratios seront rapportées.}

\item{Rec}{Un dataframe qui contient la prévision du nombre de recrues
par groupes d'espèces.}

\item{RecGaules}{Dataframe qui contient les informations sur la distribution
des gaules dans la placette.}

\item{t}{La longueur du pas de simulation en annees (en annees).}

\item{st_tot0}{Surface terrière marchande de la placette au début du pas
de simulation.}

\item{latitude}{Latitude de la placette en degrés décimal.}

\item{longitude}{Longitude de la placette en degrés décimal.}

\item{Iterj}{Itération en cours.}

\item{RandomPlacGaules}{Un dataframe contenant les effets aléatoires à
l'échelle de la placette du module de
recrutement basé sur les gaules et du module d'évolution des gaules.}

\item{Para.ratio_gaules}{Paramètres de l'équation de la prévision du ratio du
nombre de gaules par espèce.}
}
\value{
Retourne une probabilité d'absence de gaules par groupe d'espèces.
}
\description{
Fonction qui qui prévoit la proportion du nombre de gaules par espèce pour la période
de simulation suivante. Cette fonction est la première partie du modèle zero-inflated
de Rijal et al. 2023 et prévoit une probabilité d'absence de gaules par
groupe d'espèce.
}
\examples{
 #resultat <- ratio_pi_Gaules(Ratio,Rec,RecGaules,t,st_tot0,latitude,longitude,Iterj,RandomPlacGaules,Para.ratio_gaules)
print(resultat)
}
