% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recrutement_Nb_count_Gaules.R
\name{rec_count_Gaules}
\alias{rec_count_Gaules}
\title{Fonction qui prévoie le nombre de recrues par grouoe d'espèce lorsqu'il
y a présence de recrues du groupe d'espèce. Cette fonction corespond à la
deuxième portion de la fonction zero-inflated de prévision du nombre
de recrues de Rijal et al. 2023.}
\usage{
rec_count_Gaules(
  Rec,
  RecGaules,
  t,
  st_tot0,
  Iterj,
  RandomPlacGaules,
  Para.rec_gaules
)
}
\arguments{
\item{Rec}{Un dataframe qui contient la prévision du nombre de recrues par
groupes d'espèces.}

\item{RecGaules}{Dataframe qui contient les information sur la distribution
des gaules dans la placette.}

\item{t}{La longueur du pas de simulation en annees (en annees).}

\item{st_tot0}{Surface terrière marchande (DHP >9.0cm) de la placette
au début du pas de simulation.}

\item{Iterj}{Itération en cours.}

\item{RandomPlacGaules}{Un dataframe contenant les effets aléatoires à
l'échelle de la placette du module de
recrutement basé sur les gaules et du module
d'évolution des gaules.}

\item{Para.rec_gaules}{Paramètres de l'équation de prévivion du nombre de
recrues utilisant le nombre de gaules.}
}
\value{
Retourne une prévision du nombre de recrues, prévisions basées sur
         les informations provenant des recrues.
}
\description{
Fonction qui prévoie le nombre de recrues par grouoe d'espèce lorsqu'il
y a présence de recrues du groupe d'espèce. Cette fonction corespond à la
deuxième portion de la fonction zero-inflated de prévision du nombre
de recrues de Rijal et al. 2023.
}
