% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recrutement_Dhp.R
\name{rec_dhp}
\alias{rec_dhp}
\title{Fonction qui prévoit le prédicteur linéaire de l'équation du DHP des recrues}
\usage{
rec_dhp(RecSelect, st_tot0, dens_tot0, t, ntrt, Iterj, Para.rec_dhp)
}
\arguments{
\item{RecSelect}{Dataframe qui contient les préviions du nomobre de recrues par
groupe d'espèces.}

\item{st_tot0}{La surface terrière marchande de la placette en m2/ha au
début du pas de simulation.}

\item{dens_tot0}{Densitée totale de la placette en nb de tiges marchandes
(>9.0cm) par hectare au début du pas de simulation.}

\item{t}{La longueur du pas de simulation en annees (en annees).}

\item{ntrt}{Nombre de traitements de coupes partielles précédement
effectuées sur la placette.}

\item{Iterj}{Itération en cours.}

\item{Para.rec_dhp}{Un dataframe  contenant les paramettres du module
de prévision du DHP des recrues.}
}
\value{
Retourne le prédicteur linéaire de l'équation du DHP des recrues
}
\description{
Fonction qui prévoit le prédicteur linéaire de l'équation du DHP des recrues
}
\examples{
 #resultat <- rec_dhp(RecSelect,st_tot0,dens_tot0,t,ntrt,Iterj,Para.rec_dhp)
print(resultat)
}
