% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recrutement_Nb_lambda.R
\name{rec_lambda}
\alias{rec_lambda}
\title{Fonction qui prevoit le paramètre lambda de la fonction Weibull qui est
la deuxième portion du modèles zero-inflated du nombre de recrues de SaMARE.}
\usage{
rec_lambda(Rec, type_pe_Plac, st_tot0, t, Iterj, Para.rec_n)
}
\arguments{
\item{Rec}{Un dataframe qui contient la prévision du nombre de recrues par
groupes d'espèces.}

\item{type_pe_Plac}{Variable indicatrice de la taille de la placette soit
400 m2, entre 2500 et 5000 m2 inclusivement ou
une autre dimension.}

\item{st_tot0}{Surface terrière marchande (DHP >9.0cm) de la placette
au début du pas de simulation.}

\item{t}{La longueur du pas de simulation en annee (en annees).}

\item{Iterj}{Itération en cours.}

\item{Para.rec_n}{Paramètres de l'équation du nombre de recrues qui n'utilise
pas les recrues comme prédicteurs.}
}
\value{
Retourne le prédicteur linéaire du paramètre delta de la fonction
         de prévision du nombre de recrues.
}
\description{
Fonction qui prevoit le paramètre lambda de la fonction Weibull qui est
la deuxième portion du modèles zero-inflated du nombre de recrues de SaMARE.
}
