% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recrutement_Nb_pi.R
\name{rec_pi}
\alias{rec_pi}
\title{Fonction qui prévoie la probabilité de présence de recrues par grouoe d'espèce.
Cette fonction corespond à la première portion de la fonction zero-inflated
de prévision du nombre de recrues non basé sur les gaules.}
\usage{
rec_pi(Rec, t, st_tot0, ntrt, t0_aj_, type_pe_Plac, Iterj, Para.rec_n)
}
\arguments{
\item{Rec}{Un dataframe qui contient la prévision du nombre de recrues par
groupes d'espèces.}

\item{t}{La longueur du pas de simulation en annee (en annees).}

\item{st_tot0}{Surface terrière marchande (DHP >9.0cm) de la placette
au début du pas de simulation.}

\item{ntrt}{Nombre de traitements de coupes partielles précédement
effectuées sur la placette.}

\item{t0_aj_}{Temps écoulé depuis la dernière coupe partielle.}

\item{type_pe_Plac}{Variable indicatrice de la taille de la placette soit
400 m2, entre 2500 et 5000 m2 inclusivement ou
une autre dimension.}

\item{Iterj}{Itération en cours.}

\item{Para.rec_n}{Paramètres de l'équation de prévivion du nombre de
recrues sans utiliser le nombre de gaules.}
}
\value{
Retourne le prédicteur linéaire de l'équation de la prévision
         de la présence de recrues par groupe d'espèce.
}
\description{
Fonction qui prévoie la probabilité de présence de recrues par grouoe d'espèce.
Cette fonction corespond à la première portion de la fonction zero-inflated
de prévision du nombre de recrues non basé sur les gaules.
}
