% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recrutement_Produit.R
\name{rec_prod}
\alias{rec_prod}
\title{Fonction qui prevoit le prédicteur linéaire de l'équation de prévision
de la probabilité qu'une recrue de groupe d'essence feuillue ait un potentiel
de production de sciage.}
\usage{
rec_prod(RecSelect, type_pe_Plac, rid1, Iterj, Para.rec_prod)
}
\arguments{
\item{RecSelect}{Un dataframe qui contient la prévision du nombre de recrues
par groupe d'espèce.}

\item{type_pe_Plac}{Variable indicatrice de la taille de la placette soit
400 m2, soit entre 2500 et 5000 m2 inclusivement ou
soit une autre dimension.}

\item{rid1}{Variable de groupement de variables écologiques.}

\item{Iterj}{Itération en cours.}

\item{Para.rec_prod}{Paramètres de l'équation de la probabilité de présence
de potentiel sciage des recrues de groupe d'essence feuillus.}
}
\value{
Retourne le prédicteur linéaire de l'équation de prévision de la
        probabilité de présence de potentiel sciage pour les recrues de groupes
        d'essences feuillues.
}
\description{
Fonction qui prevoit le prédicteur linéaire de l'équation de prévision
de la probabilité qu'une recrue de groupe d'essence feuillue ait un potentiel
de production de sciage.
}
