% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recrutement_Vigueur.R
\name{rec_vig}
\alias{rec_vig}
\title{Fonction qui prevoit le prédicteur linéaire de l'équation de prévision
de la probabilité qu'une recrue soit vigoureuse.}
\usage{
rec_vig(RecSelect, latitude, Iterj, Para.rec_vig)
}
\arguments{
\item{RecSelect}{Un dataframe qui contient la prévision du nombre de recrues
par groupe d'espèce.}

\item{latitude}{Latitude de la placette.}

\item{Iterj}{Itération en cours.}

\item{Para.rec_vig}{Paramètres de l'équation de la probabilité que les recrues
soit vigoureuses.}
}
\value{
Retourne le prédicteur linéaire de l'équation de prévision de la
        probabilité qu'une recrue soit vigoureuse.
}
\description{
Fonction qui prevoit le prédicteur linéaire de l'équation de prévision
de la probabilité qu'une recrue soit vigoureuse.
}
\examples{
 #resultat <- rec_vig(RecSelect,latitude,Iterj,Para.rec_vig)
print(resultat)

}
