% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trouver_noms_des_colonnes_absents.R
\name{renommer_les_colonnes}
\alias{renommer_les_colonnes}
\title{Renommer les colonnes  du fichier des arbres}
\usage{
renommer_les_colonnes(data)
}
\arguments{
\item{data}{Un dataframe, représentant le fichier des arbres, dont les colonnes doivent être renommées et réorganisées.}
}
\description{
La fonction \code{renommer_les_colonnes} renomme les colonnes d'un dataframe
}
\details{
La fonction suit les étapes suivantes :
\itemize{
  \item Définir un vecteur \code{ColOrdre} contenant les noms des colonnes souhaitées dans l'ordre désiré.
  \item Convertir les noms des colonnes existantes du dataframe en minuscules pour faciliter la correspondance.
  \item Convertir les noms des colonnes souhaitées en minuscules.
  \item Pour chaque nom de colonne souhaité, chercher sa correspondance parmi les noms des colonnes existantes.
  \item Si une correspondance est trouvée, renommer la colonne existante avec le nom souhaité.
  \item Retourner le dataframe avec les noms de colonnes mis à jour.
}
}
\examples{
\dontrun{
# Supposons que nous ayons un dataframe data avec les colonnes suivantes :
# "placette", "noarbre", "espece", "etat", "dhpcm", "vigueur", "nombre",
# "sup_pe", "annee_coupe", "latitude", "longitude", "altitude", "pente",
# "ptot", "tmoy", "grw_days", "reg_eco", "type_eco", "mscr", "ntrt", "abcd"

# Appel de la fonction
# data_renomme <- renommer_les_colonnes(data)

# Le dataframe data_renomme aura les colonnes renommées et réorganisées selon ColOrdre.
}

}
