% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trouver noms des colonnes absents.R
\name{renommer_les_colonnes_gaules}
\alias{renommer_les_colonnes_gaules}
\title{Renommer les colonnes  du fichier des gaules}
\usage{
renommer_les_colonnes_gaules(data)
}
\arguments{
\item{data}{Un dataframe, représentant le fichier des gaules, dont les colonnes doivent être renommées et réorganisées.}
}
\description{
La fonction \code{renommer_les_colonnes} renomme les colonnes d'un dataframe
}
\details{
La fonction suit les étapes suivantes :
\itemize{
  \item Définir un vecteur \code{ColOrdre} contenant les noms des colonnes souhaitées dans l'ordre désiré.
  \item Convertir les noms des colonnes existantes du dataframe en minuscules pour faciliter la correspondance.
  \item Convertir les noms des colonnes souhaitées en minuscules.
  \item Pour chaque nom de colonne souhaité, chercher sa correspondance parmi les noms des colonnes existantes.
  \item Si une correspondance est trouvée, renommer la colonne existante avec le nom souhaité.
  \item Retourner le dataframe avec les noms de colonnes mis à jour.
}
}
\examples{
\dontrun{
# Supposons que nous ayons un dataframe `data` avec les colonnes suivantes :
# "placette", "espece", "grespece", "dhpcm", "nombre", "sup_pe"

# Appel de la fonction
data_renomme <- renommer_les_colonnes_gaules(data)

# Le dataframe data_renomme aura les colonnes renommées et réorganisées selon ColOrdre.
}
}
