% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaMARE.R
\name{SaMARE}
\alias{SaMARE}
\title{Fonction qui effectue la simulation de l'évolution des arbres de plusieurs placettes
avec le simulateur SaMARE pour une itération. Cette fonction effectue la simulation et
appelle chacune des fonction permettant de prévoir la mortalité, l'accroissement,
le recrutement et l'évolution des paramètres de simulation.}
\usage{
SaMARE(
  Random,
  RandomGaules,
  Data,
  Gaules,
  Iteration,
  Horizon,
  RecruesGaules,
  MCH,
  CovParms,
  CovParmsGaules,
  Para,
  ParaGaules,
  Omega,
  OmegaGaules,
  seed_value = NULL
)
}
\arguments{
\item{Random}{Un dataframe contenant des effets aléatoires à l'échelle de
la placette et de la période de simulation pour les modules
de SaMARE qui n'utilisent pas les gaules.}

\item{RandomGaules}{Un dataframe contenant des effets aléatoires à l'échelle de
la placette et de la période de simulation pour les modules
de SaMARE qui utilisent les gaules.}

\item{Data}{Un dataframe contenant une liste des arbres de dimension marchande
des placettes à simuler ainsi que des variables nécessaires à
la simulation.}

\item{Gaules}{Un dataframe contenant le nombre de gaules par classe de diamètre
et par groupe d'essence d'une placette à
simuler ainsi que des variables nécessaires à la simulation.}

\item{Iteration}{le numéro de l'itération}

\item{Horizon}{Nombre de périodes de 5 ans sur à simuler, valeur de 1 à 12 (ex: 6 pour 30 ans de simulation).}

\item{RecruesGaules}{Identification du module de recrutement à utiliser
\itemize{
   \item 0 pour utiliser le module de recrutement basé sur les arbres de dimension marchande (par défaut).
   \item 1 pour utiliser le module de recrutement basé sur l'inventaire des gaules. Le paramètre \code{Gaules} doit être aussi spécifié.
}}

\item{MCH}{Binaire. 1 = indique la présence de la maladie corticale du hêtre durant tout l'horizon de simulation.
0 = absence de la maladie (par défaut).}

\item{CovParms}{Un dataframe contenant la variance des effets aléatoires des
équations des modules de base de SaMARE (modules 1 à 9 et 17 à 19).}

\item{CovParmsGaules}{Un dataframe contenant la variance des effets aléatoires des
équations des modules de SaMARE basés sur l'information
provenant des gaules (modules 10 à 16).}

\item{Para}{Un dataframe contenant les paramètres des modules de 1 à 9 et 17 à 19
(modules de base) de SaMARE.}

\item{ParaGaules}{Un dataframe contenant les paramètres des modules 17 à 19
(modules basés sur les gaules) de SaMARE.}

\item{Omega}{Un dataframe contenant pour chaque module de base de SaMARE
(modules 1 à 9 17 à 19) le éléments du triangle inférieur de
 la matrice de variance-covariance.}

\item{OmegaGaules}{Un dataframe contenant pour chaque module basé sur les gaules
de SaMARE (modules 10 à 16) les éléments du triangle inférieur de
la matrice de variance-covariance.}

\item{seed_value}{Optionnel. La valeur du seed pour la génération de nombres aléatoires. Généralement utilisé pour les tests de la fonction.}
}
\value{
Retourne un dataframe avec une liste d'arbres vivants et mort avec leur DHP
        pour chaque étape de 5 ans de l'horizon de simulation.
}
\description{
Fonction qui effectue la simulation de l'évolution des arbres de plusieurs placettes
avec le simulateur SaMARE pour une itération. Cette fonction effectue la simulation et
appelle chacune des fonction permettant de prévoir la mortalité, l'accroissement,
le recrutement et l'évolution des paramètres de simulation.
}
