% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIM_Billonnage_ABCD_DHP.R
\name{SIMBillonnageABCD_DHP}
\alias{SIMBillonnageABCD_DHP}
\title{Fonction prévoit la répartition par produits des arbres feuillus à l'aide des nouvelles
équation de Petro régionalisés issu des travaux du CFFM de Filip Havreljuk.}
\usage{
SIMBillonnageABCD_DHP(data, ligne)
}
\arguments{
\item{ligne}{une valeur binaire égale à 1 lorsque l'on désir avoir la sortie
qui contient une ligne par produit ou 0 lorsque l'on désir avoir la sortie
qui contient une colone par produit}

\item{Data}{UUn dataframe qui contient en ligne les arbres dont on veut prévoir
les rendements en produit à l'aide du module de billonnage Petro
régionalisé.}
}
\value{
Retourne un dataframe avec l'estimation du volume par classe de produit
         pour chacun des arbres feuillus.
}
\description{
Fonction prévoit la répartition par produits des arbres feuillus à l'aide des nouvelles
équation de Petro régionalisés issu des travaux du CFFM de Filip Havreljuk.
}
