% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateurSaMARE.R
\name{SimulSaMARE}
\alias{SimulSaMARE}
\title{fonction pour faire une simulation}
\usage{
SimulSaMARE(NbIter, AnneeDep, Horizon, RecruesGaules, Data, Gaules)
}
\arguments{
\item{NbIter}{le nombre d'iteration}

\item{AnneeDep}{année de départ}

\item{Horizon}{le nombre de periode de 5 ans}

\item{RecruesGaules}{nombre de recrue}

\item{Data}{donnée de départ}

\item{Gaules}{donnée de départ pour les gaules}
}
\description{
fonction pour faire une simulation
}
