% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateurSaMARE.R
\name{SimulSaMARE}
\alias{SimulSaMARE}
\title{Fonction qui sert à appeler le simulateur SaMARE et qui fournit les données
initiales ainsi qu'un choix de paramètres pour la simulation.}
\usage{
SimulSaMARE(NbIter, Horizon, RecruesGaules, Data, Gaules, MCH = 0)
}
\arguments{
\item{NbIter}{Valeur numérique du nombre d'iterations à effectuer (ex: 30).}

\item{Horizon}{Valeur numérique du nombre de périodes de 5 ans sur lesquelles
le simulateur effectuera ses simulations (ex: 6 pour 30 ans de simulation).}

\item{RecruesGaules}{Variable prenant la valeur de "1" pour utiliser les
paramètres de recrutement basé sur l'inventaire des gaules
de la placette et de "0" pour utiliser le module de
recrutement basé sur les arbres de dimension marchande.}

\item{Data}{Un dataframe contenant les valeurs de départ pour une liste
d'arbres à simuler. Les champs: "Placette","NoArbre","Espece",
"Etat","DHPcm","Vigueur","Nombre","Sup_PE","Annee_Coupe",
"Latitude","Longitude","Altitude","Pente","Ptot","Tmoy","GrwDays",
"Reg_Eco","Type_Eco", "MSCR","ntrt" doivent être présents. Si l'information
sur certains champs n'est pas disponible, on peut le laisser vide.}

\item{Gaules}{Un dataframe contenant les valeurs de départ du nombre de gaules
par espèce et par classe de diamètre. Cette information doit être
fournie si le paramètre "RecruesGaules=1.
Les champs: "Placette","Espece","DHPcm",#' "Nombre","Sup_PE"
doivent être présents.}

\item{MCH}{Variable prenant la valeur de 1 en présence de maladie corticale du hêtre dans
la placette et 0 lorsque la maladie est absente. Lorsque la maladie corticale
est présente,la probabilité de mortalié des hêtres est estimée avec
l'équation de l'avis technique AT-SSRF 20 de la Direction de la recherche forestière.}
}
\value{
Retourne un dataframe contenant la liste des arbres, leur état, leur DHP,
        leur hauteur et leur volume pour chaque placette, chaque pas de simulation
        et chaque iteration.
}
\description{
Fonction qui sert à appeler le simulateur SaMARE et qui fournit les données
initiales ainsi qu'un choix de paramètres pour la simulation.
}
