% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sortie_Arbre_Samare.R
\name{SortieArbreSamare}
\alias{SortieArbreSamare}
\title{cette function prend en parametre un dafra qui a déja simuller}
\usage{
SortieArbreSamare(SimulHtVol)
}
\description{
cette function prend en parametre un dafra qui a déja simuller
}
