% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sortie_Billes_Samare_Colone.R
\name{SortieBillesSamareColone}
\alias{SortieBillesSamareColone}
\title{Fonction qui structure un dataframe de sortie pour lequels on rapporte en format colone
pour les essences feuillues le volume prévu pour chacune des classe de produits
issues de la classification Petro.}
\usage{
SortieBillesSamareColone(SimulHtVolBillon)
}
\arguments{
\item{SimulHtVolBillon}{Un dataframe de sortie de la fonction de billonnage.}
}
\value{
Retourne un dataframe contenant une ligne par arbre, placette, annee,
        iteration, incluant en colone le volume en dm3 de chacune des classes
        de produit Petro des arbres de groupe d'essence feuillus.
}
\description{
Fonction qui structure un dataframe de sortie pour lequels on rapporte en format colone
pour les essences feuillues le volume prévu pour chacune des classe de produits
issues de la classification Petro.
}
