% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sortie_Billes_Samare_Ligne.R
\name{SortieBillesSamareLigne}
\alias{SortieBillesSamareLigne}
\title{cette function prend en parametre un dafra qui a déja simuller et billoner}
\usage{
SortieBillesSamareLigne(SimulHtVolBillon)
}
\description{
cette function prend en parametre un dafra qui a déja simuller et billoner
}
