% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sortie_DendroIter_Samare.R
\name{SortieDendroIterSamare}
\alias{SortieDendroIterSamare}
\title{Structurer un dataframe de sortie pour les simulations SaMARE}
\usage{
SortieDendroIterSamare(SimulHtVol, simplifier = FALSE)
}
\arguments{
\item{SimulHtVol}{Un dataframe contenant les résultats de simulation pour chacune des
itérations du simulateur SaMARE. Typiquement un résultat retourné par la fonction \code{SimulSaMARE}.}

\item{simplifier}{Un booléen indiquant si les résultats doivent être simplifiés pour ne garder
que la première et la dernière année de la simulation. Par défaut, \code{FALSE}.}
}
\value{
Un dataframe contenant, pour chaque placette, groupe d'espèce, année et itération,
la surface terrière, le volume marchand brut, le diamètre moyen quadratique et la hauteur dominante.
}
\description{
La fonction \code{SortieDendroIterSamare} structure un dataframe de sortie pour lequel
on rapporte, pour chaque placette, année, groupe d'espèce et itération, le diamètre
quadratique moyen, la surface terrière, le volume et la hauteur dominante.
Cette fonction prend en paramètre un dataframe qui a déjà été simulé.
}
\examples{
\dontrun{
# Supposons que nous ayons un dataframe `SimulHtVol` contenant les résultats de simulation

# Appel de la fonction
resultats <- SortieDendroIterSamare(SimulHtVol)

# Pour simplifier les résultats
resultats_simplifies <- SortieDendroIterSamare(SimulHtVol, simplifier = TRUE)
}

}
