% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sortie_Dendro_Samare.R
\name{SortieDendroSamare}
\alias{SortieDendroSamare}
\title{Fonction qui structure un dataframe de sortie pour lequels on rapporte pour chaque
placette, annee et groupe d'espèce le diamètre quadratique moyen,
la surface terrière, le volume et la hauteur dominante.
cette function prend en parametre un daframe qui a déja été simullé}
\usage{
SortieDendroSamare(SimulHtVol, simplifier = FALSE)
}
\arguments{
\item{SimulHtVol}{Un dataframe contenant les résultats de simulation pour chacune des
iterations du simulateur SaMARE. Typiquement un résultat retourné
par la fonction "SimulSaMARE".}

\item{simplifier}{Un booléen indiquant si les résultats doivent être simplifiés pour ne garder
que la première et la dernière année de la simulation. Par défaut, \code{FALSE}.}
}
\value{
Retourne un dataframe contenant par placette, groupe d'espèce, année
         et iteration la surface terrière le volume marchand brut, le diamètre
         moyen quadratique et la hauteur dominante.
}
\description{
Fonction qui structure un dataframe de sortie pour lequels on rapporte pour chaque
placette, annee et groupe d'espèce le diamètre quadratique moyen,
la surface terrière, le volume et la hauteur dominante.
cette function prend en parametre un daframe qui a déja été simullé
}
\examples{
#resultat <- SortieDendroSamare(SimulHtVol ,simplifier=FALSE)
print(resultat)
}
