% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sortie_Dendro_Samare.R
\name{SortieDendroSamare}
\alias{SortieDendroSamare}
\title{cette function prend en parametre un dafra qui a déja simuller et billoner}
\usage{
SortieDendroSamare(Simul)
}
\description{
cette function prend en parametre un dafra qui a déja simuller et billoner
}
