% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sortie_Dendro_Samare.R
\name{SortieDendroSamare}
\alias{SortieDendroSamare}
\title{Fonction qui structure un dataframe créé par SimulSAMARE pour lequel on rapporte pour chaque
placette, annee et groupe d'espèce la moyenne des itérations (et l'écart-type) du diamètre quadratique moyen,
de la surface terrière, du volume et de la hauteur dominante, pour les vivants.
La fonction calcul aussi les accroissements survivant, mort, recrue, brut et net.}
\usage{
SortieDendroSamare(SimulHtVol, simplifier = FALSE)
}
\arguments{
\item{SimulHtVol}{un dataframe créé par  \code{SimulSAMARE}.}

\item{simplifier}{Un booléen indiquant si les résultats doivent être simplifiés pour ne garder
que la première et la dernière année de la simulation. Par défaut, \code{FALSE}.}
}
\value{
Retourne un dataframe contenant la moyenne des itérations par placette, groupe d'espèces, année
         de la surface terrière, du volume marchand brut, du diamètre moyen quadratique et de la hauteur dominante.
         Les 10 groupes d'especes de SaMARE seront présentes dans chaque placette/annee
}
\description{
Fonction qui structure un dataframe créé par SimulSAMARE pour lequel on rapporte pour chaque
placette, annee et groupe d'espèce la moyenne des itérations (et l'écart-type) du diamètre quadratique moyen,
de la surface terrière, du volume et de la hauteur dominante, pour les vivants.
La fonction calcul aussi les accroissements survivant, mort, recrue, brut et net.
}
