% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trouver_noms_des_colonnes_absents.R
\name{trouver_noms_absents}
\alias{trouver_noms_absents}
\title{Vérifier la présence des colonnes obligatoires dans le fichier d'arbres}
\usage{
trouver_noms_absents(Data)
}
\arguments{
\item{Data}{Un dataframe représentant le fichier d'arbres.}
}
\value{
Une liste des noms des colonnes manquantes.
}
\description{
La fonction \code{trouver_noms_absents} vérifie si toutes les colonnes obligatoires
sont présentes dans un dataframe représentant le fichier d'arbres. Elle retourne une liste
des noms des colonnes manquantes, le cas échéant.
}
