% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation_colonne.R
\name{valide_Annee_Coupe}
\alias{valide_Annee_Coupe}
\title{Fonction pour vérifier que les valeurs saisies dans la colonne 'Annee_Coupe' sont correctes.}
\usage{
valide_Annee_Coupe(data)
}
\arguments{
\item{data}{fichier des arbres}
}
\value{
retourne vrai ou faux s'il détecte des erreurs.

# Exemple avec un dataframe valide
valide_Annee_Coupe(data) # Devrait retourner TRUE

# Exemple avec un dataframe invalide (colonne 'Annee_Coupe' manquante)
valide_Annee_Coupe(data) # Devrait retourner FALSE

# Exemple avec un dataframe invalide (valeurs en dehors de l'intervalle)
valide_Annee_Coupe(data) # Devrait retourner FALSE

# Exemple avec un dataframe valide (toutes les valeurs de 'Annee_Coupe' sont NA pour une 'Placette')
valide_Annee_Coupe(data) # Devrait retourner TRUE
}
\description{
Fonction pour vérifier que les valeurs saisies dans la colonne 'Annee_Coupe' sont correctes.
}
