% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation_colonne.R
\name{valide_DHPcm_gaules}
\alias{valide_DHPcm_gaules}
\title{Fonction pour vérifier que les valeurs saisies dans la colonne 'DHPcm' sont correctes.}
\usage{
valide_DHPcm_gaules(data)
}
\arguments{
\item{data}{fichier des gaules}
}
\value{
retourne vrai ou faux s'il détecte des erreurs.

# Exemple avec un dataframe valide
valide_DHPcm_gaules(data) # Devrait retourner TRUE

# Exemple avec un dataframe invalide (colonne 'DHPcm' manquante)
valide_DHPcm_gaules(data) # Devrait retourner FALSE

# Exemple avec un dataframe invalide (valeurs hors de l'intervalle)
valide_DHPcm_gaules(data) # Devrait retourner FALSE

# Exemple avec un dataframe invalide (valeurs NA)
valide_DHPcm_gaules(data) # Devrait retourner FALSE
}
\description{
Fonction pour vérifier que les valeurs saisies dans la colonne 'DHPcm' sont correctes.
}
