% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation_colonne.R
\name{valide_MSCR}
\alias{valide_MSCR}
\title{Fonction pour vérifier que les valeurs saisies dans la colonne 'MSCR' sont correctes.}
\usage{
valide_MSCR(data)
}
\arguments{
\item{data}{fichier des arbres}
}
\value{
retourne vrai ou faux s'il détecte des erreurs.
}
\description{
Fonction pour vérifier que les valeurs saisies dans la colonne 'MSCR' sont correctes.
}
\examples{
# Exemple avec un dataframe valide
valide_MSCR(data) # Devrait retourner TRUE

# Exemple avec un dataframe invalide (colonne 'MSCR' manquante)
valide_MSCR(data) # Devrait retourner FALSE

# Exemple avec un dataframe invalide (valeurs non autorisées dans 'MSCR')
valide_MSCR(data) # Devrait retourner FALSE

# Exemple avec un dataframe invalide (valeurs NA dans 'MSCR' sans 'Vigueur' appropriée)
valide_MSCR(data) # Devrait retourner FALSE

}
