% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation_colonne.R
\name{valide_ntrt}
\alias{valide_ntrt}
\title{Fonction pour vérifier que les valeurs saisies dans la colonne 'ntrt' sont correctes.}
\usage{
valide_ntrt(data)
}
\arguments{
\item{data}{fichier des arbres}
}
\value{
retourne vrai ou faux s'il détecte des erreurs.
}
\description{
Fonction pour vérifier que les valeurs saisies dans la colonne 'ntrt' sont correctes.
}
\examples{
# Exemple avec un dataframe valide
valide_ntrt(data) # Devrait retourner TRUE

# Exemple avec un dataframe invalide (colonne 'ntrt' manquante)
valide_ntrt(data) # Devrait retourner FALSE

# Exemple avec un dataframe invalide (valeurs non autorisées dans 'ntrt')
valide_ntrt(data) # Devrait retourner FALSE

# Exemple avec un dataframe invalide (plusieurs valeurs différentes pour 'ntrt' par 'Placette')
valide_ntrt(data) # Devrait retourner FALSE

}
