% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation_colonne.R
\name{valide_Pente}
\alias{valide_Pente}
\title{Fonction pour vérifier que les valeurs saisies dans la colonne 'Pente' sont correctes.}
\usage{
valide_Pente(data)
}
\arguments{
\item{data}{fichier des arbres}
}
\value{
retourne vrai ou faux s'il détecte des erreurs.
}
\description{
Fonction pour vérifier que les valeurs saisies dans la colonne 'Pente' sont correctes.
}
\examples{
# Exemple avec un dataframe valide
valide_Pente(data) # Devrait retourner TRUE

# Exemple avec un dataframe invalide (colonne 'Pente' manquante)
valide_Pente(data) # Devrait retourner FALSE

# Exemple avec un dataframe invalide (valeurs en dehors de l'intervalle)
valide_Pente(data) # Devrait retourner FALSE

# Exemple avec un dataframe invalide (valeurs NA)
valide_Pente(data) # Devrait retourner FALSE

# Exemple avec un dataframe invalide (plusieurs valeurs différentes pour 'Pente' par 'Placette')
valide_Pente(data) # Devrait retourner FALSE

}
