% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation_colonne.R
\name{valide_placette_gaules}
\alias{valide_placette_gaules}
\title{Valide la correspondance des placettes entre les arbres et les gaules}
\usage{
valide_placette_gaules(data_arbre, data_gaules)
}
\arguments{
\item{data_arbre}{Un data.frame contenant les données des arbres.}

\item{data_gaules}{Un data.frame contenant les données des gaules.}
}
\value{
Une liste de messages d'erreurs. Si aucune erreur n'est trouvée, une liste vide est retournée.
}
\description{
Cette fonction vérifie que toutes les placettes présentes dans le fichier des arbres sont également présentes dans le fichier des gaules.
}
\examples{
data_arbre <- data.frame(Placette = c("TEM23APC5001", "TEM23APC5002", "TEM23APC5003"))
data_gaules <- data.frame(Placette = c("TEM23APC5001", "TEM23APC5002"))
valide_placette_gaules(data_arbre, data_gaules)
# Les erreurs seront retournées comme ceci
# 'Les placette suivantes n'ont pas de gaules : TEM23APC5003'
}
