% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation_colonne.R
\name{valide_Reg_Eco}
\alias{valide_Reg_Eco}
\title{Fonction pour vérifier que les valeurs saisies dans la colonne 'Reg_Eco' sont correctes.}
\usage{
valide_Reg_Eco(data)
}
\arguments{
\item{data}{fichier des arbres}
}
\value{
retourne vrai ou faux s'il détecte des erreurs.
# Exemple avec un dataframe valide
valide_Reg_Eco(data) # Devrait retourner TRUE

# Exemple avec un dataframe invalide (colonne 'Reg_Eco' manquante)
valide_Reg_Eco(data) # Devrait retourner FALSE

# Exemple avec un dataframe invalide (valeurs non autorisées)
valide_Reg_Eco(data) # Devrait retourner FALSE

# Exemple avec un dataframe invalide (valeurs NA)
valide_Reg_Eco(data) # Devrait retourner FALSE

# Exemple avec un dataframe invalide (plusieurs valeurs différentes pour 'Reg_Eco' par 'Placette')
valide_Reg_Eco(data) # Devrait retourner FALSE
}
\description{
Fonction pour vérifier que les valeurs saisies dans la colonne 'Reg_Eco' sont correctes.
}
