% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation_colonne.R
\name{valide_Sup_PE}
\alias{valide_Sup_PE}
\title{Fonction pour vérifier que les valeurs saisies dans la colonne 'Sup_PE' sont correctes.}
\usage{
valide_Sup_PE(data)
}
\arguments{
\item{data}{fichier des arbres}
}
\value{
retourne vrai ou faux s'il détecte des erreurs.

# Exemple avec un dataframe valide
valide_Sup_PE(data) # Devrait retourner TRUE

# Exemple avec un dataframe invalide (colonne 'Sup_PE' manquante)
valide_Sup_PE(data) # Devrait retourner FALSE

# Exemple avec un dataframe invalide (valeurs en dehors de l'intervalle)
valide_Sup_PE(data) # Devrait retourner FALSE

# Exemple avec un dataframe invalide (valeurs NA)
valide_Sup_PE(data) # Devrait retourner FALSE

# Exemple avec un dataframe invalide (plusieurs valeurs différentes pour 'Sup_PE' par 'Placette')
valide_Sup_PE(data) # Devrait retourner FALSE
}
\description{
Fonction pour vérifier que les valeurs saisies dans la colonne 'Sup_PE' sont correctes.
}
