% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation_colonne.R
\name{verifier_arbre_uniques_par_placette}
\alias{verifier_arbre_uniques_par_placette}
\title{Fonction pour vérifier que chaque arbres est unique dans chaque placette}
\usage{
verifier_arbre_uniques_par_placette(data)
}
\arguments{
\item{data}{fichier des arbres}
}
\value{
Retourne vrai ou faux s'il y a des arbres qui se répètent.
 # Exemple avec un dataframe valide
verifier_arbre_uniques_par_placette(data) # Devrait retourner TRUE

# Exemple avec un dataframe invalide (colonne 'NoArbre' manquante)
verifier_arbre_uniques_par_placette(data) # Devrait retourner FALSE

# Exemple avec un dataframe invalide (valeurs NA dans 'Placette' ou 'NoArbre')
verifier_arbre_uniques_par_placette(data) # Devrait retourner FALSE

# Exemple avec un dataframe invalide (arbres non uniques dans une placette)
verifier_arbre_uniques_par_placette(data) # Devrait retourner FALSE
}
\description{
Fonction pour vérifier que chaque arbres est unique dans chaque placette
}
