% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vigueur.R
\name{vig}
\alias{vig}
\title{Fonction qui prévoit l'évolution de la classe de vigueur des arbres.
La vigueur peut prendre 2 valeurs soit "vigoureux" ou "non-vigoureux".}
\usage{
vig(Vig, type_pe_Plac, rid1, Iterj, Para.vig)
}
\arguments{
\item{Vig}{Un dataframe qui contient les arbres pour lesquels on veut
prévoir l'évolution de la classe de vigueur des arbres.}

\item{type_pe_Plac}{Variable indicatrice de la taille de la placette soit
400 m2, soit entre 2500 et 5000 m2 inclusivement ou
une autre dimension.}

\item{rid1}{variable de groupement de variables écologiques.}

\item{Iterj}{Itération en cours.}

\item{Para.vig}{Un dataframe  contenant les parametres du module d'évolution
de la vigueur des arbres.}
}
\value{
Retourne le prédicteur linéaire de l'équation de la probabilité que
        l'arbre soit de classe "vigoureux".
}
\description{
Fonction qui prévoit l'évolution de la classe de vigueur des arbres.
La vigueur peut prendre 2 valeurs soit "vigoureux" ou "non-vigoureux".
}
